% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{antiquities_act}
\alias{antiquities_act}
\title{Trump Might Be The First President To Scrap A National Monument}
\format{A data frame with 344 rows representing acts and 9 variables:
\describe{
  \item{current_name}{Current name of piece of land designated under the Antiquities Act}
  \item{states}{State(s) or territory where land is located}
  \item{original_name}{If included, original name of piece of land designated under the Antiquities Act}
  \item{current_agency}{Current land management agency. NPS = National Parks Service, BLM = Bureau of Land Management, USFS = US Forest Service, FWS = US Fish and Wildlife Service, NOAA = National Oceanic and National Oceanic and Atmospheric Administration}
  \item{action}{Type of action taken on land}
  \item{date}{Date of action}
  \item{year}{Year of action}
  \item{pres_or_congress}{President or congress that issued action}
  \item{acres_affected}{Acres affected by action. Note that total current acreage is not included. National monuments that cover ocean are listed in square miles.} 
}}
\source{
National Parks Conservation Association \url{https://www.npca.org/} and National Parks Service Archeology Program \url{https://www.nps.gov/history/archeology/sites/antiquities/MonumentsList.htm}
}
\usage{
antiquities_act
}
\description{
The raw data behind the story
"Trump Might Be The First President To Scrap A National Monument"
\url{https://fivethirtyeight.com/features/trump-might-be-the-first-president-to-scrap-a-national-monument/}.
}
\keyword{datasets}
