% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{august_senate_polls}
\alias{august_senate_polls}
\title{How Much Trouble Is Ted Cruz Really In?}
\format{A data frame with 594 rows representing senate polls, and 11 variables:
\describe{
  \item{cycle}{the election year}
  \item{state}{the state of the poll}
  \item{senate_class}{the class of the senate}
  \item{start_date}{the start date of the poll}
  \item{end_date}{the end odate of the poll}
  \item{DEM_poll}{the percent of support for the Democrat during the poll}
  \item{REP_poll}{the percent of support for the Republican during the poll}
  \item{DEM_result}{the result percent of support for the Democrat during the election}
  \item{REP_result}{the result percent of support for the Republican during the election}
  \item{error}{the difference between the percent of support of one party during the poll and the result percent of support for the same party during the election}
  \item{absolute_error}{the absolutel value of the error value}
  }}
\source{
Emerson College’s poll of registered voters \url{https://www.emerson.edu/sites/default/files/Files/Academics/ecp-tx-aug2018-pr.pdf}
}
\usage{
august_senate_polls
}
\description{
The raw data behind the story
"How Much Trouble Is Ted Cruz Really In?"
\url{https://fivethirtyeight.com/features/how-much-trouble-is-ted-cruz-really-in/}.
}
\keyword{datasets}
