% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{avengers}
\alias{avengers}
\title{Joining The Avengers Is As Deadly As Jumping Off A Four-Story Building}
\format{A data frame with 173 rows representing characters and 21 variables:
\describe{
  \item{url}{The URL of the comic character on the Marvel Wikia}
  \item{name_alias}{The full name or alias of the character}
  \item{appearances}{The number of comic books that character appeared in as of April 30 }
  \item{current}{Is the member currently active on an avengers affiliated team?}
  \item{gender}{The recorded gender of the character}
  \item{probationary_intro}{Sometimes the character was given probationary status as an Avenger, this is the date that happened}
  \item{full_reserve_avengers_intro}{The month and year the character was introduced as a full or reserve member of the Avnegers}
  \item{year}{The year the character was introduced as a full or reserve member of the Avengers}
  \item{years_since_joining}{2015 minus the year}
  \item{honorary}{The status of the avenger, if they were given "Honorary" Avenger status, if they are simply in the "Academy," or "Full" otherwise}
  \item{death1}{TRUE if the Avenger died, FALSE if not.} 
  \item{return1}{TRUE if the Avenger returned from their first death, FALSE if they did not, blank if not applicable}
  \item{death2}{TRUE if the Avenger died a second time after their revival, FALSE if they did not, blank if not applicable}
  \item{return2}{TRUE if the Avenger returned from their second death, FALSE if they did not, blank if not applicable}
  \item{death3}{TRUE if the Avenger died a third time after their second revival, FALSE if they did not, blank if not applicable}
  \item{return3}{TRUE if the Avenger returned from their third death, FALSE if they did not, blank if not applicable}
  \item{death4}{TRUE if the Avenger died a fourth time after their third revival, FALSE if they did not, blank if not applicable}
  \item{return4}{TRUE if the Avenger returned from their fourth death, FALSE if they did not, blank if not applicable}
  \item{death5}{TRUE if the Avenger died a fifth time after their fourth revival, FALSE if they did not, blank if not applicable}
  \item{return5}{TRUE if the Avenger returned from their fifth death, FALSE if they did not, blank if not applicable}
  \item{notes}{Descriptions of deaths and resurrections.}
}}
\source{
Deaths of Marvel comic book characters between the time they joined the Avengers and April 30, 2015, the week before Secret Wars #1.
}
\usage{
avengers
}
\description{
The raw data behind the story
"Joining The Avengers Is As Deadly As Jumping Off A Four-Story Building"
\url{http://fivethirtyeight.com/features/avengers-death-comics-age-of-ultron/}.
}
\keyword{datasets}

