% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{bad_drivers}
\alias{bad_drivers}
\title{Dear Mona, Which State Has The Worst Drivers?}
\format{A data frame with 51 rows representing the 50 states + D.C. and 8 variables:
\describe{
  \item{state}{State}
  \item{num_drivers}{Number of drivers involved in fatal collisions per billion miles}
  \item{perc_speeding}{Percentage of drivers involved in fatal collisions who were speeding}
  \item{perc_alcohol}{Percentage of drivers involved in fatal collisions who were alcohol-impaired}
  \item{perc_not_distracted}{Percentage of drivers involved in fatal collisions who were not distracted}
  \item{perc_no_previous}{Percentage of drivers involved in fatal collisions who had not been involved in any previous accidents}
  \item{insurance_premiums}{Car insurance premiums ($)}
  \item{losses}{Losses incurred by insurance companies for collisions per insured driver ($)}
}}
\source{
National Highway Traffic Safety Administration 2012, National Highway
Traffic Safety Administration 2009 & 2012, National Association of Insurance
Commissioners 2010 & 2011.
}
\usage{
bad_drivers
}
\description{
The raw data behind the story
"Dear Mona, Which State Has The Worst Drivers?"
\url{http://fivethirtyeight.com/datalab/which-state-has-the-worst-drivers/}
}
\keyword{datasets}
