% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_elaine.R
\docType{data}
\name{bad_drivers_2}
\alias{bad_drivers_2}
\title{How Many Bad Drivers Are There in Different States?}
\format{A data frame with 5 rows representing senate polls, and 4 variables:
\describe{
  \item{state}{the state}
  \item{number_of_drivers_involved_in_fatal_collisions_per_million_miles}{the number of drivers involved in fatal collisions per billion miles}
  \item{percentage_of_drivers_involved_in_fatal_collisions_who_were_speeding}{the percentage Of Drivers Involved In Fatal Collisions Who Were Speeding}
  \item{percentage_of_drivers_involved_in_fatal_collisions_who_were_alcohol_impaired}{the percentage Of Drivers Involved In Fatal Collisions Who Were Alcohol-Impaired}
  }}
\source{
Emerson College’s poll of registered voters \url{https://www.emerson.edu/sites/default/files/Files/Academics/ecp-tx-aug2018-pr.pdf}
}
\usage{
bad_drivers_2
}
\description{
The raw data behind the story
"How Much Trouble Is Ted Cruz Really In?"
\url{https://fivethirtyeight.com/features/how-much-trouble-is-ted-cruz-really-in/}.
}
\keyword{datasets}
