% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R, R/data_rachel.R
\docType{data}
\name{Bad_Drivers_ops}
\alias{Bad_Drivers_ops}
\title{How Much Trouble Is Ted Cruz Really In?}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 5 columns.}
\usage{
Bad_Drivers_ops

Bad_Drivers_ops
}
\description{
The raw data behind the story
"Bad_Drivers_ops"

The raw data behind the story
"Bad_Drivers_ops"
}
\keyword{datasets}
