% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{bechdel}
\alias{bechdel}
\title{The Dollar-And-Cents Case Against Hollywood's Exclusion of Women}
\format{A data frame with 1794 rows representing movies and 15 variables:
\describe{
  \item{year}{Year of release}
  \item{imdb}{Text to append \url{http://www.imdb.com/title/} to for IMDB url. E.g. \url{http://www.imdb.com/title/tt1711425}}
  \item{title}{Movie test}
  \item{test}{bechdel test result}
  \item{clean_test}{bechdel test result}
  \item{binary}{Bechdel Test PASS vs FAIL binary}
  \item{budget}{Film budget} 
  \item{domgross}{Domestic gross}
  \item{intgross}{International gross}
  \item{code}{Bechdel Code}
  \item{budget_2013}{Budget in 2013 inflation adjusted dollars}
  \item{domgross_2013}{Domestic gross in 2013 inflation adjusted dollars} 
  \item{intgross_2013}{International gross in 2013 inflation adjusted dollars}
  \item{period_code}{}
  \item{decade_code}{}
}}
\source{
\url{www.bechdeltest.com} and \url{www.the-numbers.com}
}
\usage{
bechdel
}
\description{
The raw data behind the story
"The Dollar-And-Cents Case Against Hollywood's Exclusion of Women"
\url{http://fivethirtyeight.com/features/the-dollar-and-cents-case-against-hollywoods-exclusion-of-women/}.
}
\keyword{datasets}

