% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{cand_events_20150130}
\alias{cand_events_20150130}
\title{Who Will Run For President: Romney Is Out}
\format{A data frame with 74 rows representing events attended by potential presidential primary candidates and 8 variables:
\describe{
  \item{person}{Potential presidential candidate}
  \item{party}{Political party}
  \item{state}{State of event}
  \item{event}{Name of event}
  \item{type}{Type of event}
  \item{date}{Date of event}
  \item{link}{Link to event}
  \item{snippet}{Snippet of event description}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/potential-candidates}
}
\usage{
cand_events_20150130
}
\description{
The raw data behind the story
"Who Will Run For President: Romney Is Out"
\url{http://fivethirtyeight.com/datalab/romney-not-running-for-president/}.
}
\seealso{
\code{\link{cand_state_20150130}}, \code{\link{cand_events_20150114}}, and \code{\link{cand_state_20150114}}
}
\keyword{datasets}

