% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cand_state_20150130}
\alias{cand_state_20150130}
\title{Who Will Run For President: Romney Is Out}
\format{
A data frame with 27 rows representing potential presidential primary candidates and 5 variables:
\describe{
  \item{person}{Potential presidential candidate}
  \item{party}{Political party}
  \item{date}{Date of event}
  \item{latest}{Latest statement}
  \item{score}{Likelihood of running score, 1 = Not running, 5 = Definitely running}
}
}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/potential-candidates}
}
\usage{
cand_state_20150130
}
\description{
The raw data behind the story
"Who Will Run For President: Romney Is Out"
\url{https://fivethirtyeight.com/features/romney-not-running-for-president/}.
}
\seealso{
\code{\link{cand_events_20150130}}, \code{\link{cand_events_20150114}}, and \code{\link{cand_state_20150114}}
}
\keyword{datasets}
