% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meredith.R
\docType{data}
\name{chess_transfers}
\alias{chess_transfers}
\title{Chess Transfers}
\format{A data frame with 932 rows representing international player transfers and 5 variables:
\describe{
  \item{url}{The corresponding website on the World Chess Federation page which details the transfers of a given year.}
  \item{id}{An numeric identifier for the chess player who transfered.}
  \item{federation}{The current national federation of the chess player}
  \item{form_fed}{The national federation from which the chress player has transfered.}
  \item{transfer_date}{The date at which the transfer took place.}
}}
\source{
World Chess Federation \url{https://ratings.fide.com/fedchange.phtml}
}
\usage{
chess_transfers
}
\description{
The raw data behind the story
"American Chess Is Great Again"
\url{https://fivethirtyeight.com/features/american-chess-is-great-again/}.
}
\keyword{datasets}
