% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{classic_rock_raw_data}
\alias{classic_rock_raw_data}
\title{Why Classic Rock Isn’t What It Used To Be}
\format{A data frame with 37,673 rows representing song plays and 8 variables:
\describe{
  \item{song}{Song name}
  \item{artist}{Artist name}
  \item{callsign}{Station callsign}
  \item{time}{Time of song play in seconds elapsed since January 1, 1970}
  \item{date_time}{Time of song play in date/time format}
  \item{unique_id}{Unique ID for each song play} 
  \item{combined}{Song and artist name combined}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/classic-rock}
}
\usage{
classic_rock_raw_data
}
\description{
The raw data behind the story
"Why Classic Rock Isn’t What It Used To Be"
\url{http://fivethirtyeight.com/features/why-classic-rock-isnt-what-it-used-to-be/}.
}
\seealso{
\code{\link{classic_rock_song_list}}
}
\keyword{datasets}

