% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{classic_rock_song_list}
\alias{classic_rock_song_list}
\title{Why Classic Rock Isn't What It Used To Be}
\format{
A data frame with 2230 rows representing unique songs and 7 variables:
\describe{
  \item{song}{Song name}
  \item{artist}{Artist name}
  \item{release_year}{Release year as listed in SongFacts}
  \item{combined}{Song and artist name combined}
  \item{has_year}{Logical variable of whether release year is included}
  \item{playcount}{Number of plays across all stations}
  \item{playcount_has_year}{Number of plays across all stations if a year was found}
}
}
\source{
SongFacts and \url{https://github.com/fivethirtyeight/data/tree/master/classic-rock}
}
\usage{
classic_rock_song_list
}
\description{
The raw data behind the story
"Why Classic Rock Isn't What It Used To Be"
\url{https://fivethirtyeight.com/features/why-classic-rock-isnt-what-it-used-to-be/}.
}
\seealso{
\code{\link{classic_rock_raw_data}}
}
\keyword{datasets}
