% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{congress_age}
\alias{congress_age}
\title{Both Republicans And Democrats Have an Age Problem}
\format{A data frame with 18,635 rows representing members of Congress
(House and Senate) and 13 variables:
\describe{
  \item{congress}{Congress number.}
  \item{chamber}{Chamber of congress: House of Representatives or Senate.}
  \item{bioguide}{bioguide}
  \item{firstname}{First name}
  \item{middlename}{Middle name}
  \item{lastname}{Last name}
  \item{suffix}{Suffix}
  \item{birthday}{Birthday}
  \item{state}{State abbreviation}
  \item{party}{Party abbreviation}
  \item{incumbent}{Boolean variable of whether member was an incumbent.}
  \item{termstart}{Start date of session.}
  \item{age}{Age at start of session.}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/congress-age}
}
\usage{
congress_age
}
\description{
The raw data behind the story
"Both Republicans And Democrats Have an Age Problem"
\url{http://fivethirtyeight.com/features/both-republicans-and-democrats-have-an-age-problem/}.
}
\keyword{datasets}

