% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cousin_marriage}
\alias{cousin_marriage}
\title{How Many Americans Are Married To Their Cousins?}
\format{
A data frame with 70 rows representing countries and 2 variables:
\describe{
  \item{country}{Country}
  \item{percent}{Percent of marriages that are consanguineous}
}
}
\source{
consang.net
}
\usage{
cousin_marriage
}
\description{
The raw data behind the story
"How Many Americans Are Married To Their Cousins?"
\url{https://fivethirtyeight.com/features/how-many-americans-are-married-to-their-cousins/}.
}
\keyword{datasets}
