% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{daily_show_guests}
\alias{daily_show_guests}
\title{Every Guest Jon Stewart Ever Had On 'The Daily Show'}
\format{A data frame with 2693 rows representing guests and 5 variables:
\describe{
  \item{year}{The year the episode aired}
  \item{google_knowledge_occupation}{Their occupation or office, according to Google's Knowledge Graph or, if they're not in there, how Stewart introduced them on the program.}
  \item{show}{Air date of episode. Not unique, as some shows had more than one guest}
  \item{group}{A larger group designation for the occupation. For instance, us senators, us presidents, and former presidents are all under "politicians"}
  \item{raw_guest_list}{The person or list of people who appeared on the show, according to Wikipedia. The GoogleKnowlege_Occupation only refers to one of them in a given row.}
}}
\source{
Google Knowlege Graph, The Daily Show clip library, Wikipedia.
}
\usage{
daily_show_guests
}
\description{
The raw data behind the story
"Every Guest Jon Stewart Ever Had On 'The Daily Show'"
\url{https://fivethirtyeight.com/features/every-guest-jon-stewart-ever-had-on-the-daily-show/}.
}
\keyword{datasets}
