% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets_master}
\alias{datasets_master}
\title{Master list of all datasets}
\format{
A data frame with 9 variables:
\describe{
  \item{Data Frame Name}{Name of lazy-loaded data frame}
  \item{In fivethirtyeightdata?}{Whether the (large) dataset is in the fivethirtyeightdata package}
  \item{Article Title}{Title as it appears on FiveThirtyEight.com}
  \item{URL}{Link to article on FiveThirtyEight.com}
  \item{Author 1}{Main author}
  \item{Author 2}{Second author (if any)}
  \item{Author 3}{Third author (if any)}
  \item{Date}{Date published}
  \item{Filed Under}{Tag for article}
}
}
\usage{
datasets_master
}
\description{
All datasets included in both fivethirtyeight and fivethirtyeightdata
packages
}
\keyword{datasets}
