% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{democratic_bench}
\alias{democratic_bench}
\title{Some Democrats Who Could Step Up If Hillary Isn't Ready For Hillary}
\format{
A data frame with 67 rows representing members of the Democratic Party and 3 variables:
\describe{
  \item{candidate}{Candidate}
  \item{raised_exp}{Amount the candidate was expected to raise}
  \item{raised_act}{Amount the candidate actually raised}
}
}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/democratic-bench}.
}
\usage{
democratic_bench
}
\description{
The raw data behind the story
"Some Democrats Who Could Step Up If Hillary Isn't Ready For Hillary"
\url{https://fivethirtyeight.com/features/some-democrats-who-could-step-up-if-hillary-isnt-ready-for-hillary/}.
}
\keyword{datasets}
