% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{elasticity_by_district}
\alias{elasticity_by_district}
\title{Political Elasticity Scores}
\format{
A dataset with 435 rows representing congressional districts and 2 variables
\describe{
\item{district}{congressional district}
\item{pvi_538}{pvi}
}
}
\source{
An elasticity score measures how sensitive a state or district it is to changes in the national political environment.
}
\usage{
elasticity_by_district
}
\description{
This folder contains the data behind the story 'Election Update: The House Districts That Swing The Most (And Least) With The National Mood'
\url{https://fivethirtyeight.com/features/election-update-the-house-districts-that-swing-the-most-and-least-with-the-national-mood/}
}
\note{
The original dataset only has 2 columns: "district" and "elasticity". I separated  the "district" columns into two.
For example, in row 1 of the dataset, the original "district" = "MI-5", and I separated it into “state" = "Michigan" and "district_number" = "5".
In addition, I used the full names for all states instead of abbreviations.
}
\seealso{
\code{\link{elasticity_by_state}}
}
\keyword{datasets}
