% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elasticity_by_state}
\alias{elasticity_by_state}
\title{Political Elasticity Scores}
\format{
A dataset with 435 rows representing each state and the District of Columbia and 2 variables
\describe{
\item{state}{state}
\item{pvi_538}{pvi}
}
}
\source{
An elasticity score measures how sensitive a state or district it is to changes in the national political environment.
}
\usage{
elasticity_by_state
}
\description{
This folder contains the data behind the story 'Election Update: The House Districts That Swing The Most (And Least) With The National Mood'
\url{https://fivethirtyeight.com/features/election-update-the-house-districts-that-swing-the-most-and-least-with-the-national-mood/}
}
\note{
I used the full names for all states instead of abbreviations.
}
\seealso{
\code{\link{elasticity_by_district}}
}
\keyword{datasets}
