% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{elo_blatter}
\alias{elo_blatter}
\title{Blatter's Reign At FIFA Hasn't Helped Soccer's Poor}
\format{A data frame with 191 rows representing countries and 5 variables:
\describe{
  \item{country}{FIFA member country}
  \item{elo98}{The team's Elo in 1998}
  \item{elo15}{The team's Elo in 2015}
  \item{confederation}{Confederation to which country belongs}
  \item{gdp06}{The country's purchasing power parity GDP as of 2006}
  \item{popu06}{The country's 2006 population}
  \item{gdp_source}{Source for gdp06}
  \item{popu_source}{Source for popu06}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/elo-blatter}.
}
\usage{
elo_blatter
}
\description{
The raw data behind the story
"Blatter's Reign At FIFA Hasn't Helped Soccer's Poor"
\url{http://fivethirtyeight.com/features/blatters-reign-at-fifa-hasnt-helped-soccers-poor/}.
}
\keyword{datasets}
