% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{endorsements}
\alias{endorsements}
\title{Pols And Polls Say The Same Thing: Jeb Bush Is A Weak Front-Runner}
\format{A data frame with 109 rows representing candidates and 9 variables:
\describe{
  \item{year}{Election year}
  \item{party}{Political party}
  \item{candidate}{Candidate running in primary}
  \item{endorsement_points}{Weighted endorsements through June 30th of the year before the primary}
  \item{percentage_endorsement_points}{Percentage of total weighted endorsement points for the candidate's political party through June 30th of the year before the primary}
  \item{money_raised}{Money raised through June 30th of the year before the primary}
  \item{percentage_of_money}{Percentage of total money raised by the candidate's political party through June 30th of the year before the primary}
  \item{primary_vote_percentage}{Percentage of votes won in the primary}
  \item{won_primary}{Did the candidate win the primary?}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/endorsements-june-30}
}
\usage{
endorsements
}
\description{
The raw data behind the story
"Pols And Polls Say The Same Thing: Jeb Bush Is A Weak Front-Runner"
\url{http://fivethirtyeight.com/features/pols-and-polls-say-the-same-thing-jeb-bush-is-a-weak-front-runner/}.
This data includes something we call "endorsement points," an attempt to quantify the importance of endorsements
by weighting each one according to the position held by the endorser:
10 points for each governor, 5 points for each senator and 1 point for each
representative
}
\keyword{datasets}

