% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{endorsements_2020}
\alias{endorsements_2020}
\title{The 2020 Endorsement Primary - Which Democratic candidates are receiving the most support from prominent members of their party?}
\format{
A data frame with 1000 rows representing endorsements and 13 variables:
\describe{
  \item{date}{date of the endorsement}
  \item{position}{position of the endorser}
  \item{city}{city of the endorser}
  \item{state}{state of the endorser}
  \item{endorser}{name of the endorser}
  \item{endorsee}{name of the endorsee}
  \item{endorser_party}{party of the endorser}
  \item{source}{source link of the endorsement}
  \item{order}{order of the endorsement}
  \item{category}{category of the endorsement}
  \item{body}{body of the endorsement}
  \item{district}{district}
  \item{points}{points the endorsement counts for}
  }
}
\source{
2020 endorsement tracker.
Methodology:
\url{https://fivethirtyeight.com/methodology/how-our-presidential-endorsement-tracker-works/}.
}
\usage{
endorsements_2020
}
\description{
The raw data behind the story
"The 2020 Endorsement Primary - Which Democratic candidates are receiving the most support from prominent members of their party?"
\url{https://projects.fivethirtyeight.com/2020-endorsements/democratic-primary/}.
}
\keyword{datasets}
