% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{fandango}
\alias{fandango}
\title{Be Suspicious Of Online Movie Ratings, Especially Fandango's}
\format{A data frame with 146 rows representing movies and 23 variables:
\describe{
  \item{film}{The film in question}
  \item{year}{Year of film}
  \item{rottentomatoes}{The Rotten Tomatoes Tomatometer score  for the film }
  \item{rottentomatoes_user}{The Rotten Tomatoes user score for the film }
  \item{metacritic}{The Metacritic critic score for the film}
  \item{metacritic_user}{The Metacritic user score for the film}
  \item{imdb}{The IMDb user score for the film}
  \item{fandango_stars}{The number of stars the film had on its Fandango movie page}
  \item{fandango_ratingvalue}{The Fandango ratingValue for the film, as pulled from the HTML of each page. This is the actual average score the movie obtained. }
  \item{rt_norm}{The Rotten Tomatoes Tomatometer score  for the film , normalized to a 0 to 5 point system}
  \item{rt_user_norm}{The Rotten Tomatoes user score for the film , normalized to a 0 to 5 point system}
  \item{metacritic_norm}{The Metacritic critic score for the film, normalized to a 0 to 5 point system}
  \item{metacritic_user_nom}{The Metacritic user score for the film, normalized to a 0 to 5 point system}
  \item{imdb_norm}{The IMDb user score for the film, normalized to a 0 to 5 point system}
  \item{rt_norm_round}{The Rotten Tomatoes Tomatometer score  for the film , normalized to a 0 to 5 point system and rounded to the nearest half-star}
  \item{rt_user_norm_round}{The Rotten Tomatoes user score for the film , normalized to a 0 to 5 point system and rounded to the nearest half-star}
  \item{metacritic_norm_round}{The Metacritic critic score for the film, normalized to a 0 to 5 point system and rounded to the nearest half-star}
  \item{metacritic_user_norm_round}{The Metacritic user score for the film, normalized to a 0 to 5 point system and rounded to the nearest half-star}
  \item{imdb_norm_round}{The IMDb user score for the film, normalized to a 0 to 5 point system and rounded to the nearest half-star}
  \item{metacritic_user_vote_count}{The number of user votes the film had on Metacritic}
  \item{imdb_user_vote_count}{The number of user votes the film had on IMDb}
  \item{fandango_votes}{The number of user votes the film had on Fandango}
  \item{fandango_difference}{The difference between the presented Fandango_Stars and the actual Fandango_Ratingvalue}
}}
\source{
The data from Fandango was pulled on Aug. 24, 2015.
}
\usage{
fandango
}
\description{
The raw data behind the story
"Be Suspicious Of Online Movie Ratings, Especially Fandango's"
\url{http://fivethirtyeight.com/features/fandango-movies-ratings/}.
contains every film that has a Rotten Tomatoes rating, a RT User rating, a
Metacritic score, a Metacritic User score, and IMDb score, and at least 30
fan reviews on Fandango.
}
\keyword{datasets}
