% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivethirtyeight.R
\docType{package}
\name{fivethirtyeight}
\alias{fivethirtyeight}
\alias{fivethirtyeight-package}
\title{fivethirtyeight: Data and code behind the stories and interactives at FiveThirtyEight}
\description{
Data and code behind the stories and interactives at FiveThirtyEight
}
\section{Data sets}{

Type \code{data(package="fivethirtyeight")} to view all datasets.
}
\examples{
library(fivethirtyeight)
}
\references{
FiveThirtyEight Data GitHub repository \url{https://github.com/fivethirtyeight/data}.
}

