% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{flying}
\alias{flying}
\title{41 Percent Of Fliers Think You're Rude If You Recline Your Seat}
\format{A data frame with 1040 rows representing respondents and 27 variables:
\describe{
  \item{respondent_id}{RespondentID}
  \item{gender}{Gender}
  \item{age}{Age}
  \item{height}{Height}
  \item{children_under_18}{Do you have any children under 18?}
  \item{household_income}{Household income bracket}
  \item{education}{Education Level}
  \item{location}{Location (census region)}
  \item{frequency}{How often do you travel by plane?}
  \item{recline_frequency}{Do you ever recline your seat when you fly?}
  \item{recline_obligation}{Under normal circumstances, does a person who reclines their seat during a flight have any obligation to the person sitting behind them?}
  \item{recline_rude}{Is it rude to recline your seat on a plane?}
  \item{recline_eliminate}{Given the opportunity, would you eliminate the possibility of reclining seats on planes entirely?}
  \item{switch_seats_friends}{Is it rude to ask someone to switch seats with you in order to be closer to friends?}
  \item{switch_seats_family}{Is it rude to ask someone to switch seats with you in order to be closer to family?}
  \item{wake_up_bathroom}{Is it rude to wake a passenger up if you are trying to go to the bathroom?}
  \item{wake_up_walk}{Is it rude to wake a passenger up if you are trying to walk around?}
  \item{baby}{In general, is it rude to bring a baby on a plane?}
  \item{unruly_child}{In general, is it rude to knowingly bring unruly children on a plane?}
  \item{two_arm_rests}{In a row of three seats, who should get to use the two arm rests?}
  \item{middle_arm_rest}{In a row of two seats, who should get to use the middle arm rest?}
  \item{shade}{Who should have control over the window shade?}
  \item{unsold_seat}{Is it rude to move to an unsold seat on a plane?}
  \item{talk_stranger}{Generally speaking, is it rude to say more than a few words tothe stranger sitting next to you on a plane?}
  \item{get_up}{On a 6 hour flight from NYC to LA, how many times is it acceptable to get up if you're not in an aisle seat?}
  \item{electronics}{Have you ever used personal electronics during take off or landing in violation of a flight attendant's direction?}
  \item{smoked}{Have you ever smoked a cigarette in an airplane bathroom when it was against the rules?}
}}
\source{
SurveyMonkey survey
}
\usage{
flying
}
\description{
The raw data behind the story
"41 Percent Of Fliers Think You're Rude If You Recline Your Seat"
\url{http://fivethirtyeight.com/datalab/airplane-etiquette-recline-seat}.
}
\keyword{datasets}
