% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{food_world_cup}
\alias{food_world_cup}
\title{The FiveThirtyEight International Food Association’s 2014 World Cup}
\format{A data frame with 1373 rows representing respondents and 48 variables:
\describe{
  \item{respondent_id}{}
  \item{knowledge}{}
  \item{interest}{}
  \item{gender}{}
  \item{age}{}
  \item{household_income}{} 
  \item{education}{}
  \item{location}{}
  \item{algeria}{}
  \item{argentina}{} 
  \item{australia}{}
  \item{belgium}{}
  \item{bosnia_and_herzegovina}{}
  \item{brazil}{}
  \item{cameroon}{} 
  \item{chile}{}
  \item{china}{}
  \item{colombia}{}
  \item{costa_rica}{}
  \item{croatia}{}
  \item{cuba}{}
  \item{ecuador}{}
  \item{england}{} 
  \item{ethiopia}{}
  \item{france}{}
  \item{germany}{}
  \item{ghana}{}
  \item{greece}{}
  \item{honduras}{}
  \item{india}{}
  \item{iran}{}
  \item{ireland}{}
  \item{italy}{} 
  \item{ivory_coast}{}
  \item{japan}{}
  \item{mexico}{}
  \item{nigeria}{} 
  \item{portugal}{}
  \item{russia}{}
  \item{south_korea}{}
  \item{spain}{}
  \item{switzerland}{} 
  \item{thailand}{}
  \item{the_netherlands}{}
  \item{turkey}{} 
  \item{united_states}{}
  \item{uruguay}{}
  \item{vietnam}{}
}}
\usage{
food_world_cup
}
\description{
The raw data behind the story
"The FiveThirtyEight International Food Association’s 2014 World Cup"
\url{http://fivethirtyeight.com/features/the-fivethirtyeight-international-food-associations-2014-world-cup/}.
}
\seealso{
See \url{https://github.com/fivethirtyeight/data/tree/master/food-world-cup}
}
\keyword{datasets}

