% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{forecast_results_2018}
\alias{forecast_results_2018}
\title{How FiveThirtyEight's 2018 Midterm Forecasts Did}
\format{
A dataframe with 1518 rows representing forecast results (as of December 3, 2018) and 11 variables:
\describe{
  \item{cycle}{cycle of the election}
  \item{branch}{branch of the election}
  \item{race}{election forecast for the gubernatorial race}
  \item{forecastdate}{the date of the forecast}
  \item{version}{version of the election forecast}
  \item{dem_win_prob}{the probability of winning for the Democrat}
  \item{rep_win_prob}{the probability of winning for the Republican}
  \item{category}{the predicted political affiliation of the forecast}
  \item{democrat_won}{whether the Democrat won}
  \item{republican_won}{whether the Republican won}
  \item{uncalled}{if a race was uncalled}
  }
}
\source{
FiveThirtyEight's 2018 House Forecast \url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/house/}
}
\usage{
forecast_results_2018
}
\description{
The raw data behind the story
'How FiveThirtyEight's 2018 Midterm Forecasts Did'
\url{https://fivethirtyeight.com/features/how-fivethirtyeights-2018-midterm-forecasts-did/}
}
\keyword{datasets}
