% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foul_balls}
\alias{foul_balls}
\title{We Watched 906 Foul Balls To Find Out Where The Most Dangerous Ones}
\format{
A data frame with 906 rows representing foul balls and 7 variables:
\describe{
  \item{matchup}{the two teams that played}
  \item{game_date}{date of the most foul heavy day at each stadium}
  \item{type_of_hit}{fly, grounder, line drive, popup, batter hits self}
  \item{exit_velocity}{recorded velocity of each hit}
  \item{predicted_zone}{zone predicted the foul ball would land in guaging angles}
  \item{camera_zone}{actual zone the ball landed it confirmed by camera angles}
  \item{used_zone}{zone used for analysis}
}
}
\source{
Baseball Savant \url{https://baseballsavant.mlb.com/}.
}
\usage{
foul_balls
}
\description{
The raw data behind the story
"We Watched 906 Foul Balls To Find Out Where The Most Dangerous Ones"
\url{https://fivethirtyeight.com/features/we-watched-906-foul-balls-to-find-out-where-the-most-dangerous-ones-land/}.
}
\details{
Information on the Zones from the 538 original article: Zones 1, 2 and 3 are the areas behind home plate and the dugouts. Zones 4 and 5 make up most of the foul territory outside the baselines up until the foul pole. Zones 6 and 7 include the areas beyond the foul poles.
}
\keyword{datasets}
