% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meredith.R
\docType{data}
\name{generic_polllist}
\alias{generic_polllist}
\title{Congress Generic Ballot Polls}
\format{A data frame with 934 rows representing polls and 21 variables:
\describe{
  \item{subgroup}{}
  \item{modeldate}{}
  \item{startdate}{Start date of the poll.}
  \item{enddate}{End date of the poll.}
  \item{pollster}{The organization that conducted the poll (rather than the organization that paid for or sponsored it)}
  \item{grade}{}
  \item{samplesize}{}
  \item{population}{}
  \item{weight}{}
  \item{influence}{}
  \item{dem}{}
  \item{rep}{}
  \item{adjusted_dem}{}
  \item{adjusted_rep}{}
  \item{multiversions}{}
  \item{tracking}{}
  \item{url}{}
  \item{poll_id}{}
  \item{question_id}{}
  \item{created_date}{}
  \item{timestamp}{}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/blob/master/congress-generic-ballot/README.md}
}
\usage{
generic_polllist
}
\description{
The raw data behind the story
"Are Democrats Winning The Race For Congress?"
\url{https://projects.fivethirtyeight.com/congress-generic-ballot-polls/}.
}
\keyword{datasets}
