% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meredith.R
\docType{data}
\name{generic_topline}
\alias{generic_topline}
\title{Congress Generic Ballot Polls}
\format{A data frame with 751 rows representing polls and 9 variables:
\describe{
  \item{model_date}{}
  \item{dem_estimate}{}
  \item{dem_hi}{}
  \item{dem_lo}{}
  \item{rep_estimate}{}
  \item{rep_hi}{}
  \item{rep_lo}{}
  \item{timestamp}{}
 
}}
\source{
See \url{https://github.com/fivethirtyeight/data/blob/master/congress-generic-ballot/README.md}
}
\usage{
generic_topline
}
\description{
The raw data behind the story
"Are Democrats Winning The Race For Congress?"
\url{https://projects.fivethirtyeight.com/congress-generic-ballot-polls/}.
}
\keyword{datasets}
