% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{governor_national_forecast}
\alias{governor_national_forecast}
\title{How FiveThirtyEight's 2018 Midterm Forecasts Did}
\format{A dataframe with 150 rows and 12 variables}
\usage{
governor_national_forecast
}
\description{
The raw data behind the story 
'Forecasting the races for governor'
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/governor/}
}
\keyword{datasets}
