% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{governor_state_forecast}
\alias{governor_state_forecast}
\title{How FiveThirtyEight's 2018 Midterm Forecasts Did}
\format{A dataframe with 7743 rows representing state-level results of the classic, lite, and deluxe  gubernatorial forecasts since Oct. 11, 2018. and 12 variables
\describe{
\item{forecastdate}{date of the forecast}
\item{state}{state of the forecast}
\item{district}{district of the forecast}
\item{special}{}
\item{candidate}{name of the candidate}
\item{party}{party of the candidate}
\item{incumbent}{whether the candidate is incumbent}
\item{model}{the model of the forecast}
\item{win_probability}{the probability of the corresponding party winning}
\item{voteshare}{the voteshare of the corresponding party}
\item{p10_voteshare}{the top 10 percentile of the voteshare}
\item{p90_voteshare}{the top 00 percentile of the voteshare}
}}
\source{
FiveThirtyEight’s House, Senate And Governor Models
Methodology: 
\url{https://fivethirtyeight.com/methodology/how-fivethirtyeights-house-and-senate-models-work/}
}
\usage{
governor_state_forecast
}
\description{
The raw data behind the story 
'Forecasting the races for governor'
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/governor/}
}
\keyword{datasets}
