% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{hiphop_cand_lyrics}
\alias{hiphop_cand_lyrics}
\title{Every mention of the 2016 primary candidates in hip-hop songs}
\format{A data frame with 377 rows representing hip-hop songs referencing POTUS candidates in 2016 and 8 variables:
\describe{
  \item{candidate}{Candidate referenced} 
  \item{song}{Song name} 
  \item{artist}{Artist name} 
  \item{sentiment}{Positive, negative or neutral} 
  \item{theme}{Theme of lyric} 
  \item{album_release_date}{Date of album release} 
  \item{line}{Lyrics} 
  \item{url}{Genius link} 
}}
\source{
Genius \url{http://genius.com/}
}
\usage{
hiphop_cand_lyrics
}
\description{
The raw data behind the story
"Every mention of the 2016 primary candidates in hip-hop songs"
\url{http://projects.fivethirtyeight.com/clinton-trump-hip-hop-lyrics/}.
}
\keyword{datasets}

