% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{hist_ncaa_bball_casts}
\alias{hist_ncaa_bball_casts}
\title{The NCAA Bracket: Checking Our Work}
\format{A data frame with 253 rows representing NCAA men's basketball tournament games and 6 variables:
\describe{
  \item{year}{} 
  \item{round}{} 
  \item{favorite}{} 
  \item{underdog}{} 
  \item{favorite_prob}{} 
  \item{favorite_win}{} 
}}
\source{
See \url{http://fivethirtyeight.com/datalab/the-ncaa-bracket-checking-our-work/}
}
\usage{
hist_ncaa_bball_casts
}
\description{
The raw data behind the story
"The NCAA Bracket: Checking Our Work"
\url{http://fivethirtyeight.com/datalab/the-ncaa-bracket-checking-our-work}.
}
\keyword{datasets}

