% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hist_senate_preds}
\alias{hist_senate_preds}
\title{How The FiveThirtyEight Senate Forecast Model Works}
\format{
A data frame with 207 rows representing US state elections and 5 variables:
\describe{
  \item{state}{Election}
  \item{year}{Year of election}
  \item{candidate}{Last name}
  \item{forecast_prob}{Probability of winning election per FiveThirtyEight Election Day forecast}
  \item{result}{`Win` or `Loss`}
}
}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/forecast-methodology}
}
\usage{
hist_senate_preds
}
\description{
The raw data behind the story
"How The FiveThirtyEight Senate Forecast Model Works"
\url{https://fivethirtyeight.com/features/how-the-fivethirtyeight-senate-forecast-model-works/}.
}
\keyword{datasets}
