% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_marium.R
\docType{data}
\name{index}
\alias{index}
\title{What America’s Governors Are Talking About}
\format{A data frame with 50 rows representing the 50 U.S. states and 5 variables:
\describe{
  \item{state}{the state}
  \item{governer}{the name of the state's governer}
  \item{party}{the party of the state's governer}
  \item{filename}{the filename of the speech in the speeches folder}
  \item{url}{a link to an official/media source for the speech}
  }}
\usage{
index
}
\description{
The raw data behind the story
"What America’s Governors Are Talking About"
\url{https://fivethirtyeight.com/features/what-americas-governors-are-talking-about/}
}
\keyword{datasets}
