% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{love_actually_adj}
\alias{love_actually_adj}
\title{The Definitive Analysis Of 'Love Actually,' The Greatest Christmas Movie Of Our Time}
\format{A data frame with 14 rows representing actors and 15 variables:
\describe{
  \item{actors}{}
  \item{bill_nighy}{}
  \item{keira_knightley}{}
  \item{andrew_lincoln}{}
  \item{hugh_grant}{}
  \item{colin_firth}{}
  \item{alan_rickman}{}
  \item{heike_makatsch}{}
  \item{laura_linney}{}
  \item{emma_thompson}{}
  \item{liam_neeson}{}
  \item{kris_marshall}{}
  \item{abdul_salis}{}
  \item{martin_freeman}{}
  \item{rowan_atkinson}{}
}}
\usage{
love_actually_adj
}
\description{
The raw data behind the story
"The Definitive Analysis Of 'Love Actually,' The Greatest Christmas Movie Of Our Time"
\url{http://fivethirtyeight.com/features/some-people-are-too-superstitious-to-have-a-baby-on-friday-the-13th/}.
The adjacency matrix of which actors appear in the same scene together.
}
\seealso{
\code{\link{love_actually_appearance}}.
}
\keyword{datasets}

