% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{male_flight_attend}
\alias{male_flight_attend}
\title{Dear Mona, How Many Flight Attendants Are Men?}
\format{
A data frame with 320 rows representing job categories and 2 variables:
\describe{
  \item{job_category}{Category of job}
  \item{percentage_male}{Percentage of workforce that are male}
}
}
\source{
IPUMS 2012 \url{https://usa.ipums.org/usa/}
}
\usage{
male_flight_attend
}
\description{
The raw data behind the story
"Dear Mona, How Many Flight Attendants Are Men?"
\url{https://fivethirtyeight.com/features/dear-mona-how-many-flight-attendants-are-men/}.
}
\keyword{datasets}
