% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_natalia.R
\docType{data}
\name{media_mentions_online}
\alias{media_mentions_online}
\title{Beto O'Rourke Ignored Cable News - And It Ignored Him}
\format{A data frame with 954 rows representing weeks and 6 variables:
\describe{
  \item{date}{start date for the week of coverage}
  \item{name}{candidate's name}
  \item{matched_stories}{number of stories in that week that mention the specified candidate}
  \item{all_candidate_stories}{number of stories in that week that mention any candidate}
  \item{pct_of_all_candidate_stories}{percentage of stories in which that specific candidate is mentioned out of all stories mentioning any candidate for that week (matched_stories / all_candidate_stories)}
  \item{query}{query for Media Cloud}
}}
\source{
Two collections in the Media Cloud \url{https://mediacloud.org/} database U.S. Top Online News \url{https://sources.mediacloud.org/#/collections/58722749} and U.S. Top Digital Native News \url{https://sources.mediacloud.org/#/collections/57078150}
}
\usage{
media_mentions_online
}
\description{
The raw data behind the story
"Beto O'Rourke Ignored Cable News - And It Ignored Him"
\url{https://fivethirtyeight.com/features/beto-orourke-ignored-cable-news-and-it-ignored-him/}
}
\keyword{datasets}
