% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{mediacloud_hurricanes}
\alias{mediacloud_hurricanes}
\title{Mediacloud Hurricanes Data:
 @format A data frame with 38 rows representing dates and 5 variables:
 \describe{
   \item{date}{Date}
   \item{harvey}{The number of sentences in online news which mention Hurricane Harvey on the specified date}
   \item{irma}{The number of sentences in online news which mention Hurricane Irma}
   \item{maria}{The number of sentences in online news which mention Hurricane Maria}
   \item{jose}{The number of sentences in online news which mention Hurricane Jose}
 }
 @source Mediacloud \url{https://mediacloud.org/}
 @seealso \code{\link{mediacloud_states}}, \code{\link{mediacloud_online_news}}, 
 \code{\link{mediacloud_trump}}, \code{\link{tv_hurricanes}}, \code{\link{tv_hurricanes_by_network}}, 
 \code{\link{tv_states}}, \code{\link{google_trends}}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38 rows and 5 columns.}
\usage{
mediacloud_hurricanes
}
\description{
Mediacloud Hurricanes Data:
 @format A data frame with 38 rows representing dates and 5 variables:
 \describe{
   \item{date}{Date}
   \item{harvey}{The number of sentences in online news which mention Hurricane Harvey on the specified date}
   \item{irma}{The number of sentences in online news which mention Hurricane Irma}
   \item{maria}{The number of sentences in online news which mention Hurricane Maria}
   \item{jose}{The number of sentences in online news which mention Hurricane Jose}
 }
 @source Mediacloud \url{https://mediacloud.org/}
 @seealso \code{\link{mediacloud_states}}, \code{\link{mediacloud_online_news}}, 
 \code{\link{mediacloud_trump}}, \code{\link{tv_hurricanes}}, \code{\link{tv_hurricanes_by_network}}, 
 \code{\link{tv_states}}, \code{\link{google_trends}}
}
\keyword{datasets}
