% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{mediacloud_online_news}
\alias{mediacloud_online_news}
\title{The Media Really Started Paying Attention to Puerto Rico When Trump Did}
\format{A data frame with 49 rows representing media outlets and 2 variables:
\describe{
\item{name}{Name of media outlet source included in Media Cloud's "U.S. Top Online News" collection}
\item{url}{URL of corresponding media outlet source}
}}
\source{
Mediacloud \url{https://mediacloud.org/}
}
\usage{
mediacloud_online_news
}
\description{
The raw data behind the story
"The Media Really Started Paying Attention to Puerto Rico When Trump Did"
\url{https://fivethirtyeight.com/features/the-media-really-started-paying-attention-to-puerto-rico-when-trump-did/}:
Mediacloud Top Online News Data.
}
\seealso{
\code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_states}},
\code{\link{mediacloud_trump}}, \code{\link{tv_hurricanes}}, \code{\link{tv_hurricanes_by_network}},
\code{\link{tv_states}}, \code{\link{google_trends}}
}
\keyword{datasets}
