% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mediacloud_states}
\alias{mediacloud_states}
\title{The Media Really Started Paying Attention to Puerto Rico When Trump Did}
\format{
A data frame with 51 rows representing dates and 4 variables:
\describe{
\item{date}{Date}
\item{texas}{The number of sentences in online news which mention Texas on the specified date}
\item{puerto_rico}{The number of sentences in online news which mention Puerto Rico}
\item{florida}{The number of sentences in online news which mention Florida}
}
}
\source{
Mediacloud \url{https://mediacloud.org/}
}
\usage{
mediacloud_states
}
\description{
The raw data behind the story
"The Media Really Started Paying Attention to Puerto Rico When Trump Did"
\url{https://fivethirtyeight.com/features/the-media-really-started-paying-attention-to-puerto-rico-when-trump-did/}:
Mediacloud States Data.
}
\seealso{
\code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_online_news}},
\code{\link{mediacloud_trump}}, \code{\link{tv_hurricanes}}, \code{\link{tv_hurricanes_by_network}},
\code{\link{tv_states}}, \code{\link{google_trends}}
}
\keyword{datasets}
