% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{mediacloud_trump}
\alias{mediacloud_trump}
\title{Mediacloud Trump Data:
 @format A data frame with 51 rows representing dates and 7 variables:
 \describe{
   \item{date}{Date}
   \item{puerto_rico}{The number of headlines that mention Puerto Rico on the given date}
   \item{puerto_rico_and_trump}{The number of headlines that mention Puerto Rico and either President or Trump}
   \item{florida}{The number of headlines that mention Florida}
   \item{florida_and_trump}{The number of headlines that mention Florida and either President or Trump}
   \item{texas}{The number of headlines that mention Texas}
   \item{texas_and_trump}{The number of headlines that mention Texas and either President or Trump}
 }
 @source Mediacloud \url{https://mediacloud.org/}
 @seealso \code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_states}}, 
 \code{\link{mediacloud_online_news}},\code{\link{tv_hurricanes}}, 
 \code{\link{tv_hurricanes_by_network}}, \code{\link{tv_states}}, \code{\link{google_trends}}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 7 columns.}
\usage{
mediacloud_trump
}
\description{
Mediacloud Trump Data:
 @format A data frame with 51 rows representing dates and 7 variables:
 \describe{
   \item{date}{Date}
   \item{puerto_rico}{The number of headlines that mention Puerto Rico on the given date}
   \item{puerto_rico_and_trump}{The number of headlines that mention Puerto Rico and either President or Trump}
   \item{florida}{The number of headlines that mention Florida}
   \item{florida_and_trump}{The number of headlines that mention Florida and either President or Trump}
   \item{texas}{The number of headlines that mention Texas}
   \item{texas_and_trump}{The number of headlines that mention Texas and either President or Trump}
 }
 @source Mediacloud \url{https://mediacloud.org/}
 @seealso \code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_states}}, 
 \code{\link{mediacloud_online_news}},\code{\link{tv_hurricanes}}, 
 \code{\link{tv_hurricanes_by_network}}, \code{\link{tv_states}}, \code{\link{google_trends}}
}
\keyword{datasets}
