% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meredith.R
\docType{data}
\name{mlb_elo}
\alias{mlb_elo}
\title{MLB Elo}
\format{A data frame with 215080 rows representing Elo ratings and 26 variables:
\describe{
  \item{date}{The date of the game.}
  \item{season}{The season within which the game was played.}
  \item{neutral}{}
  \item{playoff}{}
  \item{team1}{One team that participated in the game.}
  \item{team2}{The other team that participated in the match.}
  \item{elo1_pre}{The Elo rating for team1 prior to the game.}
  \item{elo2_pre}{The Elo rating for team2 prior to the game.}
  \item{elo_prob1}{}
  \item{elo_prob2}{}
  \item{elo1_post}{The Elo rating for team1 after the game.}
  \item{elo2_post}{The Elo rating for team2 after the game.}
  \item{rating1_pre}{}
  \item{rating2_pre}{}
  \item{pitcher1}{An identifier of the pitcher}
  \item{pitcher2}{}
  \item{pitcher1_rating}{}
  \item{pitcher2_rating}{}
  \item{pitcher1_adj}{}
  \item{pitcher2_adj}{}
  \item{rating1_post}{}
  \item{rating2_post}{}
  \item{score1}{The number of runs scored by team1.}
  \item{score2}{The number of runs scored by team2.}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/blob/master/mlb-elo/README.md}
}
\usage{
mlb_elo
}
\description{
The raw data behind the stories:
"The Complete History Of MLB"
\url{https://projects.fivethirtyeight.com/complete-history-of-mlb/}
}
\details{
"MLB Predictions"
\url{https://projects.fivethirtyeight.com/2017-mlb-predictions/}.
}
\keyword{datasets}
