% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{mueller_approval_polls}
\alias{mueller_approval_polls}
\title{Both Parties Think The Mueller Report Was Fair. They Just Completely Disagree On What It Says.}
\format{A dataset with 65 rows representing every job approval poll of Robert Mueller that we could find from when Mueller was appointed as special council on May 17, 2017 through May 3, 2019 and 12 variables
\describe{
\item{start}{the start date of the poll}
\item{end}{the end date of the poll}
\item{pollster}{the name of the pollster}
\item{sample_size}{the size of the poll sample}
\item{population}{unknown}
\item{text}{the text of the poll question}
\item{approve}{the number of approval in the poll}
\item{disapprove}{the number of disapproval in the poll}
\item{unsure}{the number of unsure in the poll}
\item{approve_(republican)}{the number of approval from Republican}
\item{approve_(democrat)}{the number of approval from Democrat}
\item{url}{the url of the poll}
}}
\source{
Polls, Washington Post / ABC and Washington Post / Schar School Polls
}
\usage{
mueller_approval_polls
}
\description{
The raw data behind the story
'Both Parties Think The Mueller Report Was Fair. They Just Completely Disagree On What It Says.'
\url{https://fivethirtyeight.com/features/both-parties-think-the-mueller-report-was-fair-they-just-disagree-on-what-it-says/}
}
\keyword{datasets}
