% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{murder_2015_final}
\alias{murder_2015_final}
\title{A Handful Of Cities Are Driving 2016's Rise In Murder}
\format{
A data frame with 83 rows representing large US cities and 5 variables:
\describe{
  \item{city}{Name of city}
  \item{state}{Name of state}
  \item{murders_2014}{Total murders in 2014}
  \item{murders_2015}{Total murders in 2015}
  \item{change}{2015 - 2014}
}
}
\source{
Unknown
}
\usage{
murder_2015_final
}
\description{
The raw data behind the story
"A Handful Of Cities Are Driving 2016's Rise In Murder"
\url{https://fivethirtyeight.com/features/a-handful-of-cities-are-driving-2016s-rise-in-murders/}.
}
\keyword{datasets}
