% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{murder_2016_prelim}
\alias{murder_2016_prelim}
\title{A Handful Of Cities Are Driving 2016's Rise In Murder}
\format{A data frame with 79 rows representing large US cities and 7 variables:
\describe{
  \item{city}{Name of city} 
  \item{state}{Name of state} 
  \item{murders_2015}{Total murders in 2015} 
  \item{murders_2016}{Total murders in 2016 (as of \code{as_of} date)} 
  \item{change}{2016 - 2015} 
  \item{source}{Data source for row} 
  \item{as_of}{2016 collection date} 
}}
\source{
Listed as \code{source} variable in dataset
}
\usage{
murder_2016_prelim
}
\description{
The raw data behind the story
"A Handful Of Cities Are Driving 2016's Rise In Murder"
\url{http://fivethirtyeight.com/features/a-handful-of-cities-are-driving-2016s-rise-in-murders/}.
}
\keyword{datasets}

