% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{nba_elo}
\alias{nba_elo}
\title{Complete History Of The NBA}
\format{A data frame with 126314 rows representing games in US professional basketball history and 23 variables:
\describe{
  \item{gameorder}{Play order of game in NBA history} 
  \item{game_id}{Unique ID for each game} 
  \item{lg_id}{Which league the game was played in} 
  \item{_iscopy}{Each row of data is tied to a single team for a single game, so _iscopy flags if this game_id has already occured for the opposing team in the same matchup} 
  \item{year_id}{Season id, named based on year in which the season ended} 
  \item{date_game}{Game date} 
  \item{seasongame}{Number of game in season} 
  \item{is_playoffs}{Flag for playoff games} 
  \item{team_id}{Three letter code for team name, from Basketball Reference} 
  \item{fran_id}{Franchise id. Multiple team_ids can fall under the same fran_id due to name changes or moves. Interactive is grouped by fran_id.} 
  \item{pts}{Points scored by team} 
  \item{elo_i}{Team elo entering the game} 
  \item{elo_n}{Team elo following the game} 
  \item{win_equiv}{Equivalent number of wins in a 82-game season for a team of elo_n quality} 
  \item{opp_id}{Team id of opponent} 
  \item{opp_fran}{Franchise id of opponent} 
  \item{opp_pts}{Points scored by opponent} 
  \item{opp_elo_i}{Opponent elo entering the game} 
  \item{opp_elo_n}{Opponent elo following the game} 
  \item{game_location}{Home (H), away (A), or neutral (N)} 
  \item{game_result}{Win or loss for team in the team_id column} 
  \item{forecast}{Elo-based chances of winning for the team in the team_id column, based on elo ratings and game location} 
  \item{notes}{Additional information} 
}}
\source{
Basketball Reference \url{http://basketball-reference.com}
}
\usage{
nba_elo
}
\description{
The raw data behind the story
"Complete History Of The NBA"
\url{http://fivethirtyeight.com/interactives/the-complete-history-of-every-nba-team-by-elo}.
}
\keyword{datasets}

