% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{nba_tattoos}
\alias{nba_tattoos}
\title{Accurately Counting NBA Tattoos Isn't Easy, Even If You're Up Close}
\format{A data frame with 636 rows representing National Basketball Association players and 2 variables:
\describe{
  \item{player_name}{Name of player}
  \item{tattoos}{TRUE corresponds to player having tattoos, FALSE corresponds to not}
}}
\source{
Ethan Swan \url{http://nbatattoos.tumblr.com/}
}
\usage{
nba_tattoos
}
\description{
The raw data behind the story
"Accurately Counting NBA Tattoos Isn't Easy, Even If You're Up Close"
\url{http://fivethirtyeight.com/datalab/accurately-counting-nba-tattoos-isnt-easy-even-if-youre-up-close/}.
}
\keyword{datasets}
