% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{ncaa_w_basketball_tournaments}
\alias{ncaa_w_basketball_tournaments}
\title{The Rise And Fall Of Women's NCAA Tournament Dynasties}
\format{A dataset with 2092 rows representing every team that has participated in the NCAA Division I Women's Basketball Tournament since it began in 1982 and 19 variables
\describe{
\item{year}{the year of the game which the team participated in}
\item{school}{the school of the participating team}
\item{seed}{The `(OR)` seeding designation in 1983 notes the eight teams that played an opening-round game to become the No. 8 seed in each region.}
\item{conference}{the conference record of the team (if available)}
\item{conf_w}{number of winning in conference record}
\item{conf_l}{number of losses in conference record}
\item{conf_percent}{percent of winning in conference record}
\item{reg_w}{number of winning in regular-season record}
\item{reg_l}{number of losses in regular-season record}
\item{reg_percent}{percent of winning in regular-season record}
\item{how_qual}{Whether the school qualified with an automatic bid (by winning its conference or conference tournament) or an at-large bid.}
\item{first_home_game}{Whether the school played its first-round tournament games on its home court.}
\item{tourney_w}{number of winning in tournament record}
\item{tourney_l}{number of losses in tournament record}
\item{tourney_finish}{The round of the final game for each team. OR=opening-round loss (1983 only); 1st=first-round loss; 2nd=second-round loss; RSF=loss in the Sweet 16; RF=loss in the Elite Eight; NSF=loss in the national semifinals; N2nd=national runner-up; Champ=national champions}
\item{full_w}{number of winning in full record}
\item{full_l}{number of losses in full record}
\item{full_percent}{percent of winning in full record}
}}
\source{
NCAA
}
\usage{
ncaa_w_basketball_tournaments
}
\description{
The raw data behind the story
'The Rise And Fall Of Women's NCAA Tournament Dynasties'
\url{https://fivethirtyeight.com/features/louisiana-tech-was-the-uconn-of-the-80s/}
}
\keyword{datasets}
