% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{nfl_fandom_google}
\alias{nfl_fandom_google}
\title{How Every NFL Team’s Fans Lean Politically
  
  The raw data behind the story
  "How Every NFL Team’s Fans Lean Politically"
  \url{https://fivethirtyeight.com/features/how-every-nfl-teams-fans-lean-politically}
  
 Google Data
 @format a data frame with 207 rows representing designated market areas and 9 variables:
 \describe{
   \item{dma}{Designated Market Area}
   \item{nfl}{The percentage of search traffic in the media market region related to the NFL over the past 5 years}
   \item{nba}{The percentage of search traffic in the region related to the NBA over the past 5 years}
   \item{mlb}{The percentage of search traffic in the region related to the MLB over the past 5 years}
   \item{nascar}{The percentage of search traffic in the region related to NASCAR over the past 5 years}
   \item{cbb}{The percentage of search traffic in the region related to the CBB over the past 5 years}
   \item{cfb}{The percentage of search traffic in the region related to the CFB over the past 5 years}
   \item{trump_2016_vote}{The percentage of voters in the region who voted for Trump in the 2016 Presidential Election}
   }
  @source Google Trends \url{https://g.co/trends/5P8aa}
  @examples  
  # To convert data frame to tidy data (long) format, run:
  library(tidyverse)
  nfl_fandom_google_tidy <- gather(nfl_fandom_google, sport, search_traffic, 2:8)
  @seealso \code{\link{nfl_fandom_surveymonkey}}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 207 rows and 9 columns.}
\usage{
nfl_fandom_google
}
\description{
How Every NFL Team’s Fans Lean Politically
  
  The raw data behind the story
  "How Every NFL Team’s Fans Lean Politically"
  \url{https://fivethirtyeight.com/features/how-every-nfl-teams-fans-lean-politically}
  
 Google Data
 @format a data frame with 207 rows representing designated market areas and 9 variables:
 \describe{
   \item{dma}{Designated Market Area}
   \item{nfl}{The percentage of search traffic in the media market region related to the NFL over the past 5 years}
   \item{nba}{The percentage of search traffic in the region related to the NBA over the past 5 years}
   \item{mlb}{The percentage of search traffic in the region related to the MLB over the past 5 years}
   \item{nascar}{The percentage of search traffic in the region related to NASCAR over the past 5 years}
   \item{cbb}{The percentage of search traffic in the region related to the CBB over the past 5 years}
   \item{cfb}{The percentage of search traffic in the region related to the CFB over the past 5 years}
   \item{trump_2016_vote}{The percentage of voters in the region who voted for Trump in the 2016 Presidential Election}
   }
  @source Google Trends \url{https://g.co/trends/5P8aa}
  @examples  
  # To convert data frame to tidy data (long) format, run:
  library(tidyverse)
  nfl_fandom_google_tidy <- gather(nfl_fandom_google, sport, search_traffic, 2:8)
  @seealso \code{\link{nfl_fandom_surveymonkey}}
}
\keyword{datasets}
