% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{nfl_fandom_surveymonkey}
\alias{nfl_fandom_surveymonkey}
\title{Surveymonkey Data:}
\format{a data frame with 33 rows representing teams and 25 variables:
\describe{
\item{team}{NFL team}
\item{total_respondents}{Total number of poll respondents who ranked the given team in their top 3 favorites}
\item{asian_dem}{Number of asian, democrat poll respondents who ranked the given team in their top 3 favorites}
\item{black_dem}{Number of black, democrat poll respondents who ranked the given team in their top 3 favorites}
\item{hispanic_dem}{Number of hispanic, democrat poll respondents who ranked the given team in their top 3 favorites}
\item{other_dem}{Number of democrat poll respondents who identified their race as "other" (not asian, black, hispanic, or white) and ranked the given team in their top 3 favorites}
\item{white_dem}{Number of white, democrat poll respondents who ranked the given team in their top 3 favorites}
\item{total_dem}{Total number of democrat poll respondents who ranked the given team in their top 3 favorites}
\item{asian_ind}{Number of asian, independent poll respondents who ranked the given team in their top 3 favorites}
\item{black_ind}{Number of black, independent poll respondents who ranked the given team in their top 3 favorites}
\item{hispanic_ind}{Number of hispanic, independent poll respondents who ranked the given team in their top 3 favorites}
\item{other_ind}{Number of independent poll respondents who identified their race as "other" (not asian, black, hispanic, or white) and ranked the given team in their top 3 favorites}
\item{white_ind}{Number of white, independent poll respondents who ranked the given team in their top 3 favorites}
\item{total_ind}{Total number of independent poll respondents who ranked the given team in their top 3 favorites}
\item{asian_gop}{Number of asian, republican poll respondents who ranked the given team in their top 3 favorites}
\item{black_gop}{Number of black, republican poll respondents who ranked the given team in their top 3 favorites}
\item{hispanic_gop}{Number of hispanic, republican poll respondents who ranked the given team in their top 3 favorites}
\item{other_gop}{Number of republican poll respondents who identified their race as "other" (not asian, black, hispanic, or white) and ranked the given team in their top 3 favorites}
\item{white_gop}{Number of white, republican poll respondents who ranked the given team in their top 3 favorites}
\item{total_gop}{Total number of republican poll respondents who ranked the given team in their top 3 favorites}
\item{gop_percent}{Percent of fans (who ranked the team in their top 3 favorite NFL teams) who are republicans}
\item{dem_percent}{Percent of fans who are democrats}
\item{ind_percent}{Percent of fans who are independent}
\item{white_percent}{Percent of fans who are white}
\item{nonwhite_percent}{Percent of fans who are not white}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/nfl-fandom/NFL_fandom_data-surveymonkey.csv}
}
\usage{
nfl_fandom_surveymonkey
}
\description{
Surveymonkey Data:
}
\examples{
# To convert data frame to tidy data (long) format, run:
library(tidyverse)
nfl_fandom_surveymonkey_tidy <- nfl_fandom_surveymonkey \%>\%
gather(key= race_party, value=percent, 3:20)
}
\seealso{
\code{\link{nfl_fandom_google}}
}
\keyword{datasets}
