% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{nfl_fav_team}
\alias{nfl_fav_team}
\title{The Rams Are Dead To Me, So I Answered 3,352 Questions To Find A New NFL Team}
\format{A data frame with 32 rows representing National Football League teams and 17 variables:
\describe{
  \item{team}{Name of NFL team}
  \item{fan_relations}{Fan relations - Courtesy by players, coaches and front offices toward fans, and how well a team uses technology to reach them}
  \item{ownership}{Ownership - Honesty; loyalty to core players and the community}
  \item{players}{Players - Effort on the field, likability off it}
  \item{future_wins}{Future wins - Projected wins over next 5 seasons}
  \item{bandwagon}{Bandwagon Factor - Are the team's next 5 years likely to be better than their previous 5?}
  \item{tradition}{Tradition - Championships/division titles/wins in team's entire history}
  \item{bang_buck}{Bang for the buck - Wins per fan dollars spent}
  \item{behavior}{Behavior - Suspensions by players on team since 2007, with extra weight to transgressions vs. women}
  \item{nyc_prox}{Proximity to New York City}
  \item{stlouis_prox}{Proximity to St. Louis}
  \item{afford}{Affordability - Price of tickets, parking and concessions}
  \item{small_market}{Small Market - Size of market in terms of population, where smaller is better}
  \item{stadium_exp}{Stadium experience - Quality of venue; fan-friendliness of environment; frequency of game-day promotions}
  \item{coaching}{Coaching - Strength of on-field leadership}
  \item{uniform}{Uniform - Stylishness of uniform design, according to Uni Watch's Paul Lukas}
  \item{big_market}{Big Market - Size of market in terms of population, where bigger is better}
}}
\source{
\url{https://www.espn.com/sportsnation/teamrankings}, \url{https://www.allourideas.org/nflteampickingsample}
}
\usage{
nfl_fav_team
}
\description{
The raw data behind the story
"The Rams Are Dead To Me, So I Answered 3,352 Questions To Find A New NFL Team"
\url{https://fivethirtyeight.com/features/the-rams-are-dead-to-me-so-i-answered-3352-questions-to-find-a-new-team/}.
}
\keyword{datasets}
