% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{nfltix_div_avgprice}
\alias{nfltix_div_avgprice}
\title{Who Goes To Meaningless NFL Games And Why?}
\format{A data frame with 108 rows representing National Football League games and 3 variables:
\describe{
  \item{event}{NFL divisional game info} 
  \item{division}{NFL division} 
  \item{avg_tix_price}{Average ticket price} 
}}
\source{
StubHub \url{http://www.stubhub.com/}
}
\usage{
nfltix_div_avgprice
}
\description{
The raw data behind the story
"Who Goes To Meaningless NFL Games And Why?"
\url{http://fivethirtyeight.com/datalab/who-goes-to-meaningless-nfl-games-and-why/}.
}
\keyword{datasets}

