% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{nfltix_usa_avg}
\alias{nfltix_usa_avg}
\title{Who Goes To Meaningless NFL Games And Why?}
\format{A data frame with 32 rows representing National Football League teams and 2 variables:
\describe{
  \item{team}{Name of NFL team} 
  \item{avg_tix_price}{Average ticket price} 
}}
\source{
StubHub
}
\usage{
nfltix_usa_avg
}
\description{
The raw data behind the story
"Who Goes To Meaningless NFL Games And Why?"
\url{http://fivethirtyeight.com/datalab/who-goes-to-meaningless-nfl-games-and-why/}.
}
\keyword{datasets}

