% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{nflwr_aging_curve}
\alias{nflwr_aging_curve}
\title{The Football Hall Of Fame Has A Receiver Problem}
\format{A data frame with 24 rows representing National Football League wide receiver ages and 3 variables:
\describe{
  \item{age_from}{Beginning age}
  \item{age_to}{Ending age}
  \item{trypg_change}{Change in TRY per game from one age-year to next}
}}
\source{
Unknown
}
\usage{
nflwr_aging_curve
}
\description{
The raw data behind the story
"The Football Hall Of Fame Has A Receiver Problem"
\url{https://fivethirtyeight.com/features/the-football-hall-of-fame-has-a-receiver-problem/}.
}
\keyword{datasets}
