% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nutrition_pvalues}
\alias{nutrition_pvalues}
\title{You Can't Trust What You Read About Nutrition}
\format{
A data frame with 27716 rows representing Regression fits for p-hacking and 3 variables:
\describe{
  \item{food}{Name of food (response/dependent variable)}
  \item{characteristic}{Name of characteristic (predictor/independent variable)}
  \item{p_values}{P-value from regression fit}
}
}
\source{
See \url{https://fivethirtyeight.com/features/you-cant-trust-what-you-read-about-nutrition/}
}
\usage{
nutrition_pvalues
}
\description{
The raw data behind the story
"You Can't Trust What You Read About Nutrition"
\url{https://fivethirtyeight.com/features/you-cant-trust-what-you-read-about-nutrition/}.
}
\keyword{datasets}
