% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{partisan_lean_state}
\alias{partisan_lean_state}
\title{FiveThirtyEight's Partisan Lean}
\format{A dataset with 50 rows representing states and 2 variables 
\describe{
\item{state}{the state}
\item{pvi_538}{the pvi of the state}
}}
\source{
Partisan lean is the average difference between how a state or district votes and how the country votes overall, with 2016 presidential election results weighted 50 percent, 2012 presidential election results weighted 25 percent and results from elections for the state legislature weighted 25 percent.
}
\usage{
partisan_lean_state
}
\description{
This directory contains the data for FiveThirtyEight's partisan lean, which is used in our 
[House]
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/house}
[Senate]
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/senate}
and [Governer]
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/governor/}
forecasts.
}
\seealso{
\code{\link{partisan_lean_district}}
}
\keyword{datasets}
