% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{police_deaths}
\alias{police_deaths}
\title{The Dallas Shooting Was Among The Deadliest For Police In U.S. History}
\format{
A data frame with 22800 rows representing Police officers/dogs who lost their lives and 7 variables:
\describe{
  \item{person}{Name of person/canine who died}
  \item{cause_of_death}{Cause of death}
  \item{date}{Date of event}
  \item{year}{Year of event}
  \item{canine}{TRUE if canine, FALSE if human}
  \item{dept_name}{Name of police department}
  \item{state}{State of police department}
}
}
\source{
Officer Down Memorial Page \url{https://www.odmp.org/}
}
\usage{
police_deaths
}
\description{
The raw data behind the story
"The Dallas Shooting Was Among The Deadliest For Police In U.S. History"
\url{https://fivethirtyeight.com/features/the-dallas-shooting-was-among-the-deadliest-for-police-in-u-s-history/}.
}
\keyword{datasets}
