% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tidying.R
\docType{data}
\name{police_locals_tidy}
\alias{police_locals_tidy}
\title{Most Police Don’t Live In The Cities They Serve (Tidy Data)}
\format{A data frame with 450 rows representing combinations of race and cities and 4 variables:
\describe{
  \item{city}{U.S. city}
  \item{force_size}{Number of police officers serving that city}
  \item{race}{The race categorization corresponding as the key to \code{perc_in}.}
  \item{perc_in}{Percentage of police officers who live in a city of a given \code{race}}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/police-locals}
}
\usage{
police_locals_tidy
}
\description{
The data behind the story
"Most Police Don’t Live In The Cities They Serve"
\url{http://fivethirtyeight.com/datalab/most-police-dont-live-in-the-cities-they-serve/}.
}
\details{
This data is transformed to follow tidy data principles.


The dataset includes the cities with the 75 largest police forces, with the exception
 of Honolulu for which data is not available. All calculations are based on data from
  the U.S. Census.

The Census Bureau numbers are potentially going to differ from other counts for three reasons:
\enumerate{
   \item The census category for police officers also includes sheriffs,
    transit police and others who might not be under the same jurisdiction as a city’s police department proper. The census category won’t include private security officers.
   \item The census data is estimated from 2006 to 2010; police forces
    may have changed in size since then.
   \item There is always a margin of error in census numbers; they are
    estimates, not complete counts.
}

Note: Missing values means that there are fewer than 100 police officers of that race serving that city.
}
\keyword{datasets}
