% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chester.R
\docType{data}
\name{pres_2016_trail}
\alias{pres_2016_trail}
\title{The Last 10 Weeks Of 2016 Campaign Stops In One Handy Gif}
\format{A data frame with 177 rows representing 2016 Republican and Democratic candidate campaign trail stops and 5 variables:
\describe{
  \item{candidate}{Clinton or Trump}
  \item{date}{The date of the event}
  \item{location}{The location of the event}
  \item{lat}{Latitude of the event location}
  \item{lng}{Longitude of the event location}
}}
\source{
\url{https://hillaryspeeches.com/}, \url{https://www.conservativedailynews.com/}
}
\usage{
pres_2016_trail
}
\description{
The raw data behind the story
"The Last 10 Weeks Of 2016 Campaign Stops In One Handy Gif"
\url{https://fivethirtyeight.com/features/the-last-10-weeks-of-2016-campaign-stops-in-one-handy-gif/}.
}
\keyword{datasets}
