% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{presidential_commencement_speeches}
\alias{presidential_commencement_speeches}
\title{Sitting Presidents Give Way More Commencement Speeches Than They Used To}
\format{A data frame with 154 rows representing speeches and 8 variables:
\describe{
  \item{president}{Number of president (33 is Harry Truman, the 33rd president; 44 is Barack Obama, the 44th president)}
  \item{president_name}{Name of president}
  \item{title}{Description of commencement speech}
  \item{date}{Date speech was delivered}
  \item{city}{City where speech was delivered}
  \item{state}{state where speech was delivered}
  \item{building}{Name of building in which speech was delivered}
  \item{room}{Room in which speech was delivered}
}}
\source{
American Presidency Project, Gerhard Peters and John T. Woolley \url{http://www.presidency.ucsb.edu}
}
\usage{
presidential_commencement_speeches
}
\description{
The raw data behind the story
"Sitting Presidents Give Way More Commencement Speeches Than They Used To"
\url{http://fivethirtyeight.com/features/sitting-presidents-give-way-more-commencement-speeches-than-they-used-to/}.
}
\keyword{datasets}

