% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{pulitzer}
\alias{pulitzer}
\title{Do Pulitzers Help Newspapers Keep Readers?}
\format{A data frame with 50 rows representing newspapers and 7 variables:
\describe{
  \item{newspaper}{Newspaper}
  \item{circ2004}{Daily Circulation in 2004}
  \item{circ2013}{Daily Circulation in 2013}
  \item{pctchg_circ}{Percent change in Daily Circulation from 2004 to 2013}
  \item{num_finals1990_2003}{Number of Pulitzer Prize winners and finalists from 1990 to 2003}
  \item{num_finals2004_2014}{Number of Pulitzer Prize winners and finalists from 2004 to 2014}
  \item{num_finals1990_2014}{Number of Pulitzer Prize winners and finalists from 1990 to 2014}
}}
\source{
See \url{https://fivethirtyeight.com/features/do-pulitzers-help-newspapers-keep-readers/}
}
\usage{
pulitzer
}
\description{
The raw data behind the story
"Do Pulitzers Help Newspapers Keep Readers?"
\url{https://fivethirtyeight.com/features/do-pulitzers-help-newspapers-keep-readers/}.
}
\keyword{datasets}
