% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rana.R
\docType{data}
\name{quasi_winshares}
\alias{quasi_winshares}
\title{The Twins’ Magical Run Will Likely Go Through The Yankees. (Because Of Course.)}
\format{A data frame contains 98,796 rows representing Quasi-Win Shares(QWS), prorated to 162 team games, for every player from 1901 as of Sept. 27, 2019 and,  and 16 variables:
\describe{
  \item{name_common}{Player name}
  \item{age}{Player's age on June 30 of season}
  \item{player_id}{Player's \url{https://www.baseball-reference.com/} ID}
  \item{year_id}{Season}
  \item{team_id}{Baseball-Reference ID of team}
  \item{lg_id}{League of team}
  \item{pct_pt}{Share of total team playing time (measured by plate appearances and leverage-weighted innings) for player}
  \item{WAR162}{Wins Above Replacement per 162 team games, averaging B-R and FG WAR}
  \item{quasi_ws}{Quasi Win Shares. Three times total wins created per 162 games; generated by adding WAR162 to wins BELOW replacement (determined by playing time) and rounding to nearest whole number}
  \item{stint_id}{Order of team stints for player within season}
  \item{franch_id}{Lahman DB franchise ID for team}
  \item{prev_franch}{Previous franchise played for (NA if player debuted for team)}
  \item{year_acq}{First season for player with that franchise}
  \item{year_left}{Last season for player with that franchise}
  \item{next_franch}{Next franchise played for (NA if player retired with or is currently on team)}
  \item{P}{If played as a Pitcher}
  \item{C}{If played as a Catcher}
  \item{X1B}{If played as a First Base}
  \item{X2B}{If played as a Second Base}
  \item{X3B}{If played as a Third Base}
  \item{SS}{If played as a Shortstop}
  \item{LF}{If played as a Left Field}
  \item{CF}{If played as a Center Field}
  \item{RF}{If played as a Right Field}
  }}
\source{
What’s The Best Way To Build a Major League Baseball Team? \url{https://fivethirtyeight.com/features/whats-the-best-way-to-build-a-major-league-baseball-team/}
}
\usage{
quasi_winshares
}
\description{
The raw data behind the story 
"The Twins’ Magical Run Will Likely Go Through The Yankees. (Because Of Course.)"
\url{https://fivethirtyeight.com/features/the-yankees-always-have-the-twins-number-will-this-postseason-be-different/}
}
\keyword{datasets}
