% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{riddler_castles2}
\alias{riddler_castles2}
\title{The Battle For Riddler Nation, Round 2}
\format{
A data frame with 932 rows representing submissions and 11 variables:
\describe{
  \item{castle1}{Number of troops out of 100 send to castle 1}
  \item{castle2}{Number of troops out of 100 send to castle 2}
  \item{castle3}{Number of troops out of 100 send to castle 3}
  \item{castle4}{Number of troops out of 100 send to castle 4}
  \item{castle5}{Number of troops out of 100 send to castle 5}
  \item{castle6}{Number of troops out of 100 send to castle 6}
  \item{castle7}{Number of troops out of 100 send to castle 7}
  \item{castle8}{Number of troops out of 100 send to castle 8}
  \item{castle9}{Number of troops out of 100 send to castle 9}
  \item{castle10}{Number of troops out of 100 send to castle 10}
  \item{reason}{Why did you choose your troop deployment?}
}
}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/riddler-castles}
}
\usage{
riddler_castles2
}
\description{
The raw data behind the story
"The Battle For Riddler Nation, Round 2"
\url{https://fivethirtyeight.com/features/the-battle-for-riddler-nation-round-2/}.
Analysis of the submitted solutions can be found at: \url{https://fivethirtyeight.com/features/how-much-should-you-bid-for-that-painting/}
}
\examples{
# To convert data frame to tidy data (long) format, run
library(dplyr)
library(tidyr)
library(stringr)
riddler_castles_tidy<-riddler_castles2 \%>\%
   pivot_longer(castle1:castle10, names_to = "castle" , values_to = "soldiers") \%>\%
   mutate(castle = as.numeric(str_replace(castle, "castle","")))
}
\seealso{
\code{\link{riddler_castles}}
}
\keyword{datasets}
