% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{riddler_pick_lowest}
\alias{riddler_pick_lowest}
\title{Pick A Number, Any Number}
\format{A data frame with 3660 rows representing dates and 1 variable:
\describe{
\item{your_number}{Guessed number}
\item{show_your_work}{People showing their work}
}}
\usage{
riddler_pick_lowest
}
\description{
The raw data behind the story
"Pick A Number, Any Number"
\url{https://fivethirtyeight.com/features/pick-a-number-any-number/}
}
\keyword{datasets}
