% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{san_andreas}
\alias{san_andreas}
\title{The Rock Isn't Alone: Lots Of People Are Worried About 'The Big One'}
\format{A data frame with 1013 rows representing respondents and 11 variables:
\describe{
  \item{worry_general}{In general, how worried are you about earthquakes?}
  \item{worry_bigone}{How worried are you about the "Big One," a massive, catastrophic earthquake?}
  \item{will_occur}{Do you think the "Big One" will occur in your lifetime?}
  \item{experience}{Have you ever experienced an earthquake?}
  \item{prepared }{Have you or anyone in your household taken any precautions for an earthquake (packed an earthquake survival kit, prepared an evacuation plan, etc.)?}
  \item{fam_san_andreas}{How familiar are you with the San Andreas Fault line?}
  \item{fam_yellowstone}{How familiar are you with the Yellowstone Supervolcano?}
  \item{age}{Age}
  \item{female}{Gender}
  \item{hhold_income}{How much total combined money did all members of your HOUSEHOLD earn last year?}
  \item{region}{US Region}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/san-andreas}
}
\usage{
san_andreas
}
\description{
The raw data behind the story
"The Rock Isn’t Alone: Lots Of People Are Worried About 'The Big One'"
\url{http://fivethirtyeight.com/datalab/the-rock-isnt-alone-lots-of-people-are-worried-about-the-big-one/}.
}
\keyword{datasets}
