% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{senate_national_forecast}
\alias{senate_national_forecast}
\title{Senate Forecast 2018}
\format{A dataframe with 450 rows representing national-level results of the classic, lite, and deluxe   Senate forecasts since Aug. 1, 2018 and 11 variables
\describe{
\item{forecastdate}{date of the forecast}
\item{party}{the party of the forecast}
\item{model}{the model of the forecast}
\item{win_probability}{the probability of the corresponding party winning}
\item{mean_seats}{the mean of the number of seats}
\item{median_seats}{the median number of seats}
\item{p10_seats}{the top 10 percentile of number of seats}
\item{p90_seats}{the top 90 percentile of number of seats}
\item{margin}{unknown}
\item{p10_margin}{the margin of p10_seats}
\item{p90_margin}{the margin of p90_seats}
}}
\source{
FiveThirtyEight’s House, Senate And Governor Models
Methodology:
\url{https://fivethirtyeight.com/methodology/how-fivethirtyeights-house-and-senate-models-work/}
}
\usage{
senate_national_forecast
}
\description{
This file contains links to the data behind FiveThirtyEight's 'Senate forecasts'
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/senate/}
}
\note{
The original dataset included a meaningless column called "state", and all variables under this column was "US". So this column was removed.
}
\seealso{
\code{\link{senate_seat_forecast}}
}
\keyword{datasets}
