% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{senate_seat_forecast}
\alias{senate_seat_forecast}
\title{Senate Forecast 2018}
\format{A dataframe with 28353 rows representing seat-level results of the classic, lite, and deluxe Senate forecasts since Aug. 1, 2018 and 12 variables
\describe{
\item{forecastdate}{date of the forecast}
\item{state}{state of the forecast}
\item{class}{class of the forecast}
\item{special}{unknown}
\item{candidate}{name of the candidate}
\item{party}{party of the candidate}
\item{incumbent}{whether the candidate is incumbent}
\item{model}{the model of the forecast}
\item{win_probability}{the probability of the corresponding party winning}
\item{voteshare}{the voteshare of the corresponding party}
\item{p10_voteshare}{the top 10 percentile of the voteshare}
\item{p90_voteshare}{the top 00 percentile of the voteshare}
}}
\source{
FiveThirtyEight’s House, Senate And Governor Models
Methodology:
\url{https://fivethirtyeight.com/methodology/how-fivethirtyeights-house-and-senate-models-work/}
}
\usage{
senate_seat_forecast
}
\description{
This file contains links to the data behind FiveThirtyEight's 'Senate forecasts'
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/senate/}
}
\seealso{
\code{\link{senate_national_forecast}}
}
\keyword{datasets}
