% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meredith.R
\docType{data}
\name{spi_global_rankings}
\alias{spi_global_rankings}
\title{Current SPI ratings and rankings for men's club teams}
\format{A data frame with 453 rows representing soccer rankings and 7 variables:
\describe{
  \item{name}{The name of the soccer club.}
  \item{league}{The name of the league to which the club belongs.}
  \item{rank}{A club's current global ranking.}
  \item{prev_rank}{A club's previous global ranking}
  \item{off}{Offensive rating for a given team (the higher the value the stronger the team's offense).}
  \item{def}{Defensive rating for a given team (the lower the value the stronger the team's defense).}
  \item{spi}{A club's SPI score.}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/blob/master/soccer-spi/README.md}
}
\usage{
spi_global_rankings
}
\description{
The raw data behind the stories
"Club Soccer Predictions"
\url{https://projects.fivethirtyeight.com/soccer-predictions/}
}
\details{
"Global Club Soccer Rankings"
\url{https://projects.fivethirtyeight.com/global-club-soccer-rankings/}.
}
\seealso{
\code{\link{spi_matches}}
}
\keyword{datasets}
