% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{steak_survey}
\alias{steak_survey}
\title{How Americans Like Their Steak}
\format{A data frame with 550 rows representing respondents and 15 variables:
\describe{
}}
\source{
See \url{http://fivethirtyeight.com/datalab/how-americans-like-their-steak/}
}
\usage{
steak_survey
}
\description{
The raw data behind the story
"How Americans Like Their Steak"
\url{http://fivethirtyeight.com/datalab/how-americans-like-their-steak/}.
}
\keyword{datasets}

