% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{tennis_events_time}
\alias{tennis_events_time}
\title{Why Some Tennis Matches Take Forever}
\format{
A data frame with 205 rows representing tournaments and 5 variables:
\describe{
  \item{tournament}{Name of event}
  \item{surface}{Court surface used at the event}
  \item{sec_added}{Seconds added per point for this event on this surface in years shown, from regression model controlling for players, year and other factors}
  \item{year_start}{Start year for data used from this tournament in regression}
  \item{year_end}{End year for data used from this tournament in regression}
}
}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/tennis-time}
}
\usage{
tennis_events_time
}
\description{
The raw data behind the story
"Why Some Tennis Matches Take Forever"
\url{https://fivethirtyeight.com/features/why-some-tennis-matches-take-forever/}.
}
\seealso{
\code{\link{tennis_players_time}} and \code{\link{tennis_serve_time}}
}
\keyword{datasets}
