% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{tennis_players_time}
\alias{tennis_players_time}
\title{Why Some Tennis Matches Take Forever}
\format{A data frame with 218 rows representing players and 2 variables:
\describe{
  \item{player}{Player Name}
  \item{sec_added}{Weighted average of seconds added per point as loser and winner of matches, 1991-2015, from regression model controlling for tournament, surface, year and other factors}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/tennis-time}
}
\usage{
tennis_players_time
}
\description{
The raw data behind the story
"Why Some Tennis Matches Take Forever"
\url{https://fivethirtyeight.com/features/why-some-tennis-matches-take-forever/}.
}
\seealso{
\code{\link{tennis_events_time}} and \code{\link{tennis_serve_time}}
}
\keyword{datasets}
