% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{tennis_serve_time}
\alias{tennis_serve_time}
\title{Why Some Tennis Matches Take Forever}
\format{A data frame with 120 rows representing serves and 7 variables:
\describe{
  \item{server}{Name of player serving at 2015 French Open}
  \item{sec_between}{Time in seconds between end of marked point and next serve, timed by stopwatch app}
  \item{opponent}{Opponent, receiving serve}
  \item{game_score}{Score in the current game during the timed interval between points}
  \item{set}{Set number, out of five}
  \item{game}{Score in games within the set}
  \item{date}{Date}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/tennis-time}
}
\usage{
tennis_serve_time
}
\description{
The raw data behind the story
"Why Some Tennis Matches Take Forever"
\url{http://fivethirtyeight.com/features/why-some-tennis-matches-take-forever/}.
}
\seealso{
\code{\link{tennis_events_time}} and \code{\link{tennis_players_time}}
}
\keyword{datasets}
