% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenth_circuit}
\alias{tenth_circuit}
\title{For A Trump Nominee, Neil Gorsuch’s Record Is Surprisingly Moderate On Immigration}
\format{
A data frame with 954 rows representing cases and 13 variables:
\describe{
  \item{title}{Name of the case}
  \item{date}{Date of decision}
  \item{federalreporter_cit}{Case citation, as listed in the Federal Reporter Series}
  \item{westlaw_cit}{Case citation, Westlaw format}
  \item{issue}{Issue number, in cases divided into multiple issues}
  \item{weight}{Weight per issue (total weight per case equals one)}
  \item{judge1}{Name of first judge}
  \item{judge2}{Name of second judge}
  \item{judge3}{Name of third judge}
  \item{vote1_liberal}{Vote of first judge. 1 = liberal, 0 = conservative.}
  \item{vote2_liberal}{Vote of second judge. 1 = liberal, 0 = conservative.}
  \item{vote3_liberal}{Vote of third judge. 1 = liberal, 0 = conservative.}
  \item{category}{Category of case, immigration or discrimination}
  }
}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/tenth-circuit}
}
\usage{
tenth_circuit
}
\description{
The raw data behind the story
"For A Trump Nominee, Neil Gorsuch’s Record Is Surprisingly Moderate On Immigration"
\url{https://fivethirtyeight.com/features/for-a-trump-nominee-neil-gorsuchs-record-is-surprisingly-moderate-on-immigration/}.
}
\note{
In immigration cases, partial relief to immigration petitioner is coded as liberal because the petitioner typically seeks just one core remedy (e.g., withholding of removal, adjustment of status, or asylum); in discrimination cases, partial relief is coded as multiple issues because the plaintiff often seeks separate remedies under multiple claims (e.g., disparate treatment, retaliation, etc.) and different sources of law.
}
\keyword{datasets}
