% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{trump_news}
\alias{trump_news}
\title{How Trump Hacked The Media}
\format{A data frame with 286 rows representing lead stories and 3 variables:
\describe{
  \item{date}{Date of lead story about Donald Trump.}
  \item{major_cat}{Story classification}
  \item{detail}{}
}}
\source{
Memeorandum \url{http://www.memeorandum.com/}.
}
\usage{
trump_news
}
\description{
The raw data behind the story
"How Trump Hacked The Media"
\url{https://fivethirtyeight.com/features/how-donald-trump-hacked-the-media/}.
}
\keyword{datasets}
