% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{trumpworld_polls}
\alias{trumpworld_polls}
\title{Trump World Polls Dataset:
 @format A data frame with 32 rows representng years and 40 variables: 
 \describe{
   \item{year}{Year the poll was conducted}
   \item{avg}{The average percentage people who answered the poll question positively (support the president or have a favorable view of the U.S.)}
   \item{canada}{The percentage of people from Canada who answered the poll question positively}
   \item{france}{The percentage of people from  who answered the poll question positively}
   \item{germany}{The percentage of people from  who answered the poll question positively}
   \item{greece}{The percentage of people from  who answered the poll question positively}
   \item{hungary}{The percentage of people from  who answered the poll question positively}
   \item{italy}{The percentage of people from  who answered the poll question positively}
   \item{netherlands}{The percentage of people from  who answered the poll question positively}
   \item{poland}{The percentage of people from  who answered the poll question positively}
   \item{spain}{The percentage of people from  who answered the poll question positively}
   \item{sweden}{The percentage of people from  who answered the poll question positively}
   \item{uk}{The percentage of people from  who answered the poll question positively}
   \item{russia}{The percentage of people from  who answered the poll question positively}
   \item{australia}{The percentage of people from  who answered the poll question positively}
   \item{india}{The percentage of people from  who answered the poll question positively}
   \item{indonesia}{The percentage of people from  who answered the poll question positively}
   \item{japan}{The percentage of people from  who answered the poll question positively}
   \item{philippines}{The percentage of people from  who answered the poll question positively}
   \item{south_korea}{The percentage of people from  who answered the poll question positively}
   \item{vietnam}{The percentage of people from  who answered the poll question positively}
   \item{israel}{The percentage of people from  who answered the poll question positively}
   \item{jordan}{The percentage of people from  who answered the poll question positively}
   \item{lebanon}{The percentage of people from  who answered the poll question positively}
   \item{tunisia}{The percentage of people from  who answered the poll question positively}
   \item{turkey}{The percentage of people from  who answered the poll question positively}
   \item{ghana}{The percentage of people from  who answered the poll question positively}
   \item{kenya}{The percentage of people from  who answered the poll question positively}
   \item{nigeria}{The percentage of people from  who answered the poll question positively}
   \item{senegal}{The percentage of people from  who answered the poll question positively}
   \item{south_africa}{The percentage of people from  who answered the poll question positively}
   \item{tanzania}{The percentage of people from  who answered the poll question positively}
   \item{argentina}{The percentage of people from  who answered the poll question positively}
   \item{brazil}{The percentage of people from  who answered the poll question positively}
   \item{chile}{The percentage of people from  who answered the poll question positively}
   \item{colombia}{The percentage of people from  who answered the poll question positively}
   \item{mexico}{The percentage of people from  who answered the poll question positively}
   \item{peru}{The percentage of people from  who answered the poll question positively}
   \item{venezuela}{The percentage of people from  who answered the poll question positively}
   \item{question}{The item being polled. Specificially, whether respondents:
             Have a favorable view of the U.S.
             Trust the U.S. President when it comes to foreign affairs}
  }
  @source Pew Research Center \url{http://www.pewresearch.org/fact-tank/2017/07/17/9-charts-on-how-the-world-sees-trump/}
  @examples 
  # To convert data frame to tidy data (long) format, run:
  library(tidyverse)
  trumpworld_polls_tidy <- gather(trumpworld_polls, country, percent_positive, 3:39)
  @seelalso \code{\link{trumpworld_issues}}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 40 columns.}
\usage{
trumpworld_polls
}
\description{
Trump World Polls Dataset:
 @format A data frame with 32 rows representng years and 40 variables: 
 \describe{
   \item{year}{Year the poll was conducted}
   \item{avg}{The average percentage people who answered the poll question positively (support the president or have a favorable view of the U.S.)}
   \item{canada}{The percentage of people from Canada who answered the poll question positively}
   \item{france}{The percentage of people from  who answered the poll question positively}
   \item{germany}{The percentage of people from  who answered the poll question positively}
   \item{greece}{The percentage of people from  who answered the poll question positively}
   \item{hungary}{The percentage of people from  who answered the poll question positively}
   \item{italy}{The percentage of people from  who answered the poll question positively}
   \item{netherlands}{The percentage of people from  who answered the poll question positively}
   \item{poland}{The percentage of people from  who answered the poll question positively}
   \item{spain}{The percentage of people from  who answered the poll question positively}
   \item{sweden}{The percentage of people from  who answered the poll question positively}
   \item{uk}{The percentage of people from  who answered the poll question positively}
   \item{russia}{The percentage of people from  who answered the poll question positively}
   \item{australia}{The percentage of people from  who answered the poll question positively}
   \item{india}{The percentage of people from  who answered the poll question positively}
   \item{indonesia}{The percentage of people from  who answered the poll question positively}
   \item{japan}{The percentage of people from  who answered the poll question positively}
   \item{philippines}{The percentage of people from  who answered the poll question positively}
   \item{south_korea}{The percentage of people from  who answered the poll question positively}
   \item{vietnam}{The percentage of people from  who answered the poll question positively}
   \item{israel}{The percentage of people from  who answered the poll question positively}
   \item{jordan}{The percentage of people from  who answered the poll question positively}
   \item{lebanon}{The percentage of people from  who answered the poll question positively}
   \item{tunisia}{The percentage of people from  who answered the poll question positively}
   \item{turkey}{The percentage of people from  who answered the poll question positively}
   \item{ghana}{The percentage of people from  who answered the poll question positively}
   \item{kenya}{The percentage of people from  who answered the poll question positively}
   \item{nigeria}{The percentage of people from  who answered the poll question positively}
   \item{senegal}{The percentage of people from  who answered the poll question positively}
   \item{south_africa}{The percentage of people from  who answered the poll question positively}
   \item{tanzania}{The percentage of people from  who answered the poll question positively}
   \item{argentina}{The percentage of people from  who answered the poll question positively}
   \item{brazil}{The percentage of people from  who answered the poll question positively}
   \item{chile}{The percentage of people from  who answered the poll question positively}
   \item{colombia}{The percentage of people from  who answered the poll question positively}
   \item{mexico}{The percentage of people from  who answered the poll question positively}
   \item{peru}{The percentage of people from  who answered the poll question positively}
   \item{venezuela}{The percentage of people from  who answered the poll question positively}
   \item{question}{The item being polled. Specificially, whether respondents:
             Have a favorable view of the U.S.
             Trust the U.S. President when it comes to foreign affairs}
  }
  @source Pew Research Center \url{http://www.pewresearch.org/fact-tank/2017/07/17/9-charts-on-how-the-world-sees-trump/}
  @examples 
  # To convert data frame to tidy data (long) format, run:
  library(tidyverse)
  trumpworld_polls_tidy <- gather(trumpworld_polls, country, percent_positive, 3:39)
  @seelalso \code{\link{trumpworld_issues}}
}
\keyword{datasets}
