% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tv_hurricanes}
\alias{tv_hurricanes}
\title{The Media Really Started Paying Attention to Puerto Rico When Trump Did}
\format{
A data frame with 37 rows representing dates and 5 variables:
\describe{
\item{date}{Date}
\item{harvey}{The percent of sentences in TV news that mention Hurricane Harvey on the given date}
\item{irma}{The percent of sentences in TV news that mention Hurricane Irma}
\item{maria}{The percent of sentences in TV news that mention Hurricane Maria}
\item{jose}{The percent of sentences in TV news that mention Hurricane Irma}
}
}
\source{
Internet TV News Archive \url{https://archive.org/details/tv} and
Television Explorer \url{https://television.gdeltproject.org/cgi-bin/iatv_ftxtsearch/iatv_ftxtsearch}
}
\usage{
tv_hurricanes
}
\description{
The raw data behind the story
"The Media Really Started Paying Attention to Puerto Rico When Trump Did"
\url{https://fivethirtyeight.com/features/the-media-really-started-paying-attention-to-puerto-rico-when-trump-did/}:
TV Hurricanes Data.
}
\seealso{
\code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_states}},
\code{\link{mediacloud_online_news}}, \code{\link{mediacloud_trump}},
\code{\link{tv_hurricanes_by_network}}, \code{\link{tv_states}}, \code{\link{google_trends}}
}
\keyword{datasets}
