% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{tv_hurricanes_by_network}
\alias{tv_hurricanes_by_network}
\title{TV Hurricanes by Network Data:}
\format{A data frame with 84 rows representing dates and 6 variables:
\describe{
\item{date}{Date}
\item{query}{The hurricane in question}
\item{bbc_news}{The percent of sentences on the BBC News TV channel on the given date that mention the hurricane in question}
\item{cnn}{The percent of sentences on CNN News that mention the hurricane in question}
\item{fox_news}{The percent of sentences on Fox News that mention the hurricane in question}
\item{msnbc}{The percent of sentences on MSNBC News that mention the hurricane in question}
}}
\source{
Internet TV News Archive \url{https://archive.org/details/tv} and
Television Explorer \url{https://television.gdeltproject.org/cgi-bin/iatv_ftxtsearch/iatv_ftxtsearch}
}
\usage{
tv_hurricanes_by_network
}
\description{
TV Hurricanes by Network Data:
}
\seealso{
\code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_states}}, \code{\link{mediacloud_online_news}},
\code{\link{mediacloud_trump}}, \code{\link{tv_hurricanes}}, \code{\link{tv_states}}, \code{\link{google_trends}}
}
\keyword{datasets}
