% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{tv_states}
\alias{tv_states}
\title{The Media Really Started Paying Attention to Puerto Rico When Trump Did}
\format{A data frame with 52 rows representing dates and 4 variables:
\describe{
\item{date}{Date}
\item{florida}{The percent of sentences in TV News on the given day that mention Florida}
\item{texas}{The percent of sentences in TV News on the given day that mention Texas}
\item{puerto_rico}{The percent of sentences in TV News on the given day that mention Puerto Rico}
}}
\source{
Internet TV News Archive \url{https://archive.org/details/tv} and
Television Explorer \url{https://television.gdeltproject.org/cgi-bin/iatv_ftxtsearch/iatv_ftxtsearch}
}
\usage{
tv_states
}
\description{
The raw data behind the story
"The Media Really Started Paying Attention to Puerto Rico When Trump Did"
\url{https://fivethirtyeight.com/features/the-media-really-started-paying-attention-to-puerto-rico-when-trump-did/}:
TV States Data.
}
\seealso{
\code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_states}}, \code{\link{mediacloud_online_news}},
\code{\link{mediacloud_trump}}, \code{\link{tv_hurricanes}}, \code{\link{tv_hurricanes_by_network}}, \code{\link{google_trends}}
}
\keyword{datasets}
