% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{twitter_presidents}
\alias{twitter_presidents}
\title{The Worst Tweeter in Politics Isn't Trump
 
 The raw data behind:
 "The Worst Tweeter in Politics Isn't Trump"
 \url{https://fivethirtyeight.com/features/the-worst-tweeter-in-politics-isnt-trump/}
 
 Presidents Dataset:
 @format A data frame with 6439 rows describing individual tweets and 8 variables:
 \describe{
   \item{created_at}{The date and time the tweet was posted}
   \item{text}{The text of the tweet}
   \item{url}{The link to the tweet}
   \item{replies}{The number of replies to the tweet}
   \item{retweets}{The number of retweets}
   \item{favorites}{The number of favorites}
   \item{user}{The user posting the tweet}
   \item{president}{The name of the president who posted the tweet}
 }
 @details Data on President Obama's tweets collected on Oct 20, President Trump's tweets collected on Oct 23.
 @source Twitter \url(https://twitter.com/BarackObama) and
 \url{https://twitter.com/realDonaldTrump}
 @seealso \code{\link{senators}}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6439 rows and 8 columns.}
\usage{
twitter_presidents
}
\description{
The Worst Tweeter in Politics Isn't Trump
 
 The raw data behind:
 "The Worst Tweeter in Politics Isn't Trump"
 \url{https://fivethirtyeight.com/features/the-worst-tweeter-in-politics-isnt-trump/}
 
 Presidents Dataset:
 @format A data frame with 6439 rows describing individual tweets and 8 variables:
 \describe{
   \item{created_at}{The date and time the tweet was posted}
   \item{text}{The text of the tweet}
   \item{url}{The link to the tweet}
   \item{replies}{The number of replies to the tweet}
   \item{retweets}{The number of retweets}
   \item{favorites}{The number of favorites}
   \item{user}{The user posting the tweet}
   \item{president}{The name of the president who posted the tweet}
 }
 @details Data on President Obama's tweets collected on Oct 20, President Trump's tweets collected on Oct 23.
 @source Twitter \url(https://twitter.com/BarackObama) and
 \url{https://twitter.com/realDonaldTrump}
 @seealso \code{\link{senators}}
}
\keyword{datasets}
