% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{undefeated}
\alias{undefeated}
\title{Mayweather Is Defined By The Zero Next To His Name
 
 The raw data behind: 
 "Mayweather Is Defined By The Zero Next To His Name"
 \url{https://fivethirtyeight.com/features/mayweather-is-defined-by-the-zero-next-to-his-name/}
 @format A data frame with 2125 rows representing boxing matches and 4 variables:
 \describe{
   \item{name}{Name of boxer}
   \item{url}{URL with the boxer's record}
   \item{date}{Date of the match}
   \item{wins}{Number of cumulative wins for the boxer including the match at the specified date}
 }
 @source Box Rec \url{http://boxrec.com/}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2125 rows and 4 columns.}
\usage{
undefeated
}
\description{
Mayweather Is Defined By The Zero Next To His Name
 
 The raw data behind: 
 "Mayweather Is Defined By The Zero Next To His Name"
 \url{https://fivethirtyeight.com/features/mayweather-is-defined-by-the-zero-next-to-his-name/}
 @format A data frame with 2125 rows representing boxing matches and 4 variables:
 \describe{
   \item{name}{Name of boxer}
   \item{url}{URL with the boxer's record}
   \item{date}{Date of the match}
   \item{wins}{Number of cumulative wins for the boxer including the match at the specified date}
 }
 @source Box Rec \url{http://boxrec.com/}
}
\keyword{datasets}
