% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{US_births_2000_2014}
\alias{US_births_2000_2014}
\title{Some People Are Too Superstitious To Have A Baby On Friday The 13th}
\format{A data frame with 5479 rows representing dates and 6 variables:
\describe{
  \item{year}{Year}
  \item{month}{Month}
  \item{date_of_month}{Day}
  \item{date}{POSIX date}
  \item{day_of_week}{Abbreviation of day of week}
  \item{births}{Number of births}
}}
\source{
Social Security Administration
}
\usage{
US_births_2000_2014
}
\description{
The raw data behind the story
"Some People Are Too Superstitious To Have A Baby On Friday The 13th"
\url{https://fivethirtyeight.com/features/some-people-are-too-superstitious-to-have-a-baby-on-friday-the-13th/}.
}
\seealso{
\code{\link{US_births_1994_2003}}.
}
\keyword{datasets}
