% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{weather_check}
\alias{weather_check}
\title{Where People Go To Check The Weather}
\format{A data frame with 928 rows representing respondents and 9 variables:
\describe{
  \item{respondent_id}{Respondent ID}
  \item{ck_weather}{Do you typically check a daily weather report?}
  \item{weather_source}{How do you typically check the weather?}
  \item{weather_source_site}{If they responded "A specific website or app" when asked how they typically check the weather, they were asked to write-in the app or website they used.}
  \item{ck_weather_watch}{If you had a smartwatch (like the soon to be released Apple Watch), how likely or unlikely would you be to check the weather on that device?}
  \item{age}{Age}
  \item{female}{Gender}
  \item{hhold_income}{How much total combined money did all members of your HOUSEHOLD earn last year?}
  \item{region}{US Region}
}}
\source{
The source of the data is a Survey Monkey Audience poll commissioned by FiveThirtyEight and conducted from April 6 to April 10, 2015.  See \url{https://github.com/fivethirtyeight/data/tree/master/weather-check}
}
\usage{
weather_check
}
\description{
The raw data behind the story
"Where People Go To Check The Weather"
\url{http://fivethirtyeight.com/datalab/weather-forecast-news-app-habits/}.
}
\keyword{datasets}

