% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_marium.R
\docType{data}
\name{words}
\alias{words}
\title{What America’s Governors Are Talking About}
\format{A data frame with 2,223 rows representing phrases and 9 variables:
\describe{
  \item{phrase}{one-, two-, and three-word phrases spoken repeatedly}
  \item{category}{thematic categories for the phrases}
  \item{d_speeches}{number of Democratic speeches containing the phrase}
  \item{r_speeches}{number of Republican speeches containing the phrase}
  \item{total}{total number of speeches containing the phrase}
  \item{percent_of_d_speeches}{percent of the 23 Democratic speeches containing the phrase}
  \item{percent_of_r_speeches}{percent of the 27 Republican speeches containing the phrase}
  \item{chi2}{the chi-square test statistic for statistical significance}
  \item{pval}{p-value for chi^2 test}
  }}
\source{
The chi-square test statistic \url{https://scikit-learn.org/stable/modules/generated/sklearn.feature_selection.chi2.html#sklearn.feature_selection.chi2}
}
\usage{
words
}
\description{
The raw data behind the story
"What America’s Governors Are Talking About"
\url{https://fivethirtyeight.com/features/what-americas-governors-are-talking-about/}
}
\keyword{datasets}
