% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airline_safety}
\alias{airline_safety}
\title{Should Travelers Avoid Flying Airlines That Have Had Crashes in the Past?}
\format{
A data frame with 56 rows representing airlines and 9 variables:
\describe{
  \item{airline}{airline}
  \item{incl_reg_subsidiaries}{indicates that regional subsidiaries are included}
  \item{avail_seat_km_per_week}{available seat kilometers flown every week}
  \item{incidents_85_99}{Total number of incidents, 1985-1999}
  \item{fatal_accidents_85_99}{Total number of fatal accidents, 1985-1999}
  \item{fatalities_85_99}{Total number of fatalities, 1985-1999}
  \item{incidents_00_14}{Total number of incidents, 2000-2014}
  \item{fatal_accidents_00_14}{Total number of fatal accidents, 2000-2014}
  \item{fatalities_00_14}{Total number of fatalities, 2000-2014}
}
}
\source{
Aviation Safety Network \url{https://aviation-safety.net}.
}
\usage{
airline_safety
}
\description{
The raw data behind the story
"Should Travelers Avoid Flying Airlines That Have Had Crashes in the Past?"
\url{https://fivethirtyeight.com/features/should-travelers-avoid-flying-airlines-that-have-had-crashes-in-the-past/}.
}
\examples{
# To convert data frame to tidy data (long) format, run:
library(dplyr)
library(tidyr)
library(stringr)
airline_safety_tidy <- airline_safety \%>\%
  pivot_longer(-c(airline, incl_reg_subsidiaries, avail_seat_km_per_week), 
    names_to = "type", values_to = "count") \%>\%
  mutate(
    period = str_sub(type, start=-5),
    period = str_replace_all(period, "_", "-"),
    type = str_sub(type, end=-7)
  )
}
\keyword{datasets}
