% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{bechdel}
\alias{bechdel}
\title{The Dollar-And-Cents Case Against Hollywood’s Exclusion of Women}
\format{A data frame with 1794 rows representing movies and 15 variables:
\describe{
  \item{year}{}
  \item{imdb}{}
  \item{title}{}
  \item{test}{}
  \item{clean_test}{}
  \item{binary}{}
  \item{budget}{} 
  \item{domgross}{}
  \item{intgross}{}
  \item{code}{}
  \item{budget_2013}{}
  \item{domgross_2013}{} 
  \item{intgross_2013}{}
  \item{period_code}{}
  \item{decade_code}{}
}}
\source{
\url{www.bechdeltest.com} and \url{www.the-numbers.com}
}
\usage{
bechdel
}
\description{
The raw data behind the story
"The Dollar-And-Cents Case Against Hollywood’s Exclusion of Women"
\url{http://fivethirtyeight.com/features/the-dollar-and-cents-case-against-hollywoods-exclusion-of-women/}.
}
\keyword{datasets}

