% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{bechdel}
\alias{bechdel}
\title{The Dollar-And-Cents Case Against Hollywood's Exclusion of Women}
\format{A data frame with 1794 rows representing movies and 15 variables:
\describe{
  \item{year}{Year of release}
  \item{imdb}{Text to construct IMDB url. Ex: \url{https://www.imdb.com/title/tt1711425}}
  \item{title}{Movie test}
  \item{test}{bechdel test result (detailed, with discrepancies indicated)}
  \item{clean_test}{bechdel test result (detailed): \code{ok} = passes test, \code{dubious}, \code{men} = women only talk about men, \code{notalk} = women don't talk to each other, \code{nowomen} = fewer than two women }
  \item{binary}{Bechdel Test PASS vs FAIL binary}
  \item{budget}{Film budget}
  \item{domgross}{Domestic (US) gross}
  \item{intgross}{Total International (i.e., worldwide) gross}
  \item{code}{Bechdel Code}
  \item{budget_2013}{Budget in 2013 inflation adjusted dollars}
  \item{domgross_2013}{Domestic gross (US) in 2013 inflation adjusted dollars}
  \item{intgross_2013}{Total International (i.e., worldwide) gross in 2013 inflation adjusted dollars}
  \item{period_code}{}
  \item{decade_code}{}
}}
\source{
\url{http://bechdeltest.com/} and \url{https://www.the-numbers.com/}. The original data can be found at \url{https://github.com/fivethirtyeight/data/tree/master/bechdel}.
}
\usage{
bechdel
}
\description{
The raw data behind the story
"The Dollar-And-Cents Case Against Hollywood's Exclusion of Women"
\url{https://fivethirtyeight.com/features/the-dollar-and-cents-case-against-hollywoods-exclusion-of-women/}.
}
\details{
A vignette of an analysis of this dataset using the \code{tidyverse} can be found on \href{https://CRAN.R-project.org/package=fivethirtyeight/vignettes/bechdel.html}{CRAN} or by running:
\code{vignette("bechdel", package = "fivethirtyeight")}
}
\keyword{datasets}
