% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biopics}
\alias{biopics}
\title{'Straight Outta Compton' Is The Rare Biopic Not About White Dudes}
\format{
A data frame with 761 rows representing movies and 14 variables:
\describe{
  \item{title}{Title of the film.}
  \item{site}{Text to construct IMDB url. Ex: \url{https://www.imdb.com/title/tt1711425}}
  \item{country}{Country of origin.}
  \item{year_release}{Year of release.}
  \item{box_office}{Gross earnings at U.S. box office.}
  \item{director}{Director of film.}
  \item{number_of_subjects}{The number of subjects featured in the film.}
  \item{subject}{The actual name of the featured subject.}
  \item{type_of_subject}{The occupation of subject or reason for recognition.}
  \item{race_known}{Indicates whether the subject's race was discernible based on background of self, parent, or grandparent.}
  \item{subject_race}{Race of the subject.}
  \item{person_of_color}{Dummy variable that indicates person of color.}
  \item{subject_sex}{Sex of subject.}
  \item{lead_actor_actress}{The actor or actress who played the subject.}
}
}
\source{
IMDB \url{https://www.imdb.com/}
}
\usage{
biopics
}
\description{
The raw data behind the story
"'Straight Outta Compton' Is The Rare Biopic Not About White Dudes"
\url{https://fivethirtyeight.com/features/straight-outta-compton-is-the-rare-biopic-not-about-white-dudes/}.
An analysis using this data was contributed by Pradeep Adhokshaja as a package vignette at
\url{https://fivethirtyeight-r.netlify.com/articles/biopics.html}.
}
\keyword{datasets}
