% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lizette.R
\docType{data}
\name{cabinet_turnover}
\alias{cabinet_turnover}
\title{Two Years In, Turnover In Trump’s Cabinet Is Still Historically High}
\format{
A data frame with 312 rows representing cabinet members and 8 variables: 
\describe{
  \item{president}{Surname of of sitting President}
  \item{position}{Cabinet Position}
  \item{appointee}{Appointee's full name}
  \item{start}{Date the appointee was sworn in}
  \item{end}{Date the appointee left office}
  \item{length}{Length of Tenure, in days}
  \item{days}{Days into administration that the appointee left office}
  \item{combined}{Whether or not Cabinet member served in more than one administrations}
  }
}
\source{
from Federal Government Websites and News Reports
}
\usage{
cabinet_turnover
}
\description{
The raw data behind the story 
"Two Years In, Turnover In Trump’s Cabinet Is Still Historically High"
\url{https://fivethirtyeight.com/features/two-years-in-turnover-in-trumps-cabinet-is-still-historically-high/}.
}
\keyword{datasets}
