% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{cand_events_20150114}
\alias{cand_events_20150114}
\title{Looking For Clues: Who Is Going To Run For President In 2016?}
\format{A data frame with 42 rows representing events attended in Iowa and New Hampshire by potential presidential primary candidates and 8 variables:
\describe{
  \item{person}{Potential presidential candidate}
  \item{party}{Political party}
  \item{state}{State of event}
  \item{event}{Name of event}
  \item{type}{Type of event}
  \item{date}{Date of event}
  \item{link}{Link to event}
  \item{snippet}{Snippet of event description}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/tree/master/potential-candidates}
}
\usage{
cand_events_20150114
}
\description{
The raw data behind the story
"Looking For Clues: Who Is Going To Run For President In 2016?"
\url{https://fivethirtyeight.com/features/2016-president-who-is-going-to-run/}.
}
\seealso{
\code{\link{cand_state_20150114}}, \code{\link{cand_events_20150130}}, and \code{\link{cand_state_20150130}}
}
\keyword{datasets}
