% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivethirtyeight.R
\docType{package}
\name{fivethirtyeight}
\alias{fivethirtyeight}
\title{fivethirtyeight: Data and Code Behind the Stories and Interactives at 'FiveThirtyEight'}
\description{
An R package that provides access to the code and data sets
published by FiveThirtyEight \url{https://github.com/fivethirtyeight/data}. Note
that while we received guidance from editors at 538, this package is not
officially published by 538. Contribute to this package at
\url{https://github.com/rudeboybert/fivethirtyeight}.
}
\examples{
# Example usage:
library(fivethirtyeight)
\donttest{head(bechdel)}

# All information about any data set can be found in the help file:
\donttest{?bechdel}

# To view a list of all data sets:
\donttest{data(package = "fivethirtyeight")}

# To view a detailed list of all data sets:
\donttest{vignette("fivethirtyeight", package = "fivethirtyeight")}
}
