% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_albert.R
\docType{data}
\name{food_world_cup}
\alias{food_world_cup}
\title{The FiveThirtyEight International Food Association's 2014 World Cup}
\format{A data frame with 1373 rows representing respondents and 48 variables:
\describe{
  \item{respondent_id}{Respondent ID}
  \item{knowledge}{Generally speaking, how would you rate your level of knowledge of cuisines from different parts of the world?}
  \item{interest}{How much, if at all, are you interested in cuisines from different parts of the world?}
  \item{gender}{Gender}
  \item{age}{Age}
  \item{household_income}{Household income bracket}
  \item{education}{Education Level}
  \item{location}{Location (census region)}
  \item{algeria}{Cuisine of Algeria}
  \item{argentina}{Cuisine of Argentina} 
  \item{australia}{Cuisine of Australia}
  \item{belgium}{Cuisine of Belgium}
  \item{bosnia_and_herzegovina}{Cuisine of Bosnia & Herzegovina}
  \item{brazil}{Cuisine of Brazil}
  \item{cameroon}{Cuisine of Cameroon} 
  \item{chile}{Cuisine of Chile}
  \item{china}{Cuisine of China}
  \item{colombia}{Cuisine of Colombia}
  \item{costa_rica}{Cuisine of Costa Rica}
  \item{croatia}{Cuisine of Croatia}
  \item{cuba}{Cuisine of Cuba}
  \item{ecuador}{Cuisine of Ecuador}
  \item{england}{Cuisine of England} 
  \item{ethiopia}{Cuisine of Ethiopia}
  \item{france}{Cuisine of France}
  \item{germany}{Cuisine of Germany}
  \item{ghana}{Cuisine of Ghana}
  \item{greece}{Cuisine of Greece}
  \item{honduras}{Cuisine of Honduras}
  \item{india}{Cuisine of India}
  \item{iran}{Cuisine of Iran}
  \item{ireland}{Cuisine of Ireland}
  \item{italy}{Cuisine of Italy} 
  \item{ivory_coast}{Cuisine of Ivory Coast}
  \item{japan}{Cuisine of Japan}
  \item{mexico}{Cuisine of Mexico}
  \item{nigeria}{Cuisine of Nigeria} 
  \item{portugal}{Cuisine of Portugal}
  \item{russia}{Cuisine of Russia}
  \item{south_korea}{Cuisine of South Korea}
  \item{spain}{Cuisine of Spain}
  \item{switzerland}{Cuisine of Switzerland} 
  \item{thailand}{Cuisine of Thailand}
  \item{the_netherlands}{Cuisine of the Netherlands}
  \item{turkey}{Cuisine of Turkey} 
  \item{united_states}{Cuisine of the United States}
  \item{uruguay}{Cuisine of Uruguay}
  \item{vietnam}{Cuisine of Vietnam}
}}
\usage{
food_world_cup
}
\description{
The raw data behind the story
"The FiveThirtyEight International Food Association's 2014 World Cup"
\url{http://fivethirtyeight.com/features/the-fivethirtyeight-international-food-associations-2014-world-cup/}.
For all the countries below, the response to the following question is presented:
"Please rate how much you like the traditional cuisine of X"
\itemize{
 \item{5: }{I love this country's traditional cuisine. I think it's one of the best in the world.}
 \item{4: }{I like this country's traditional cuisine. I think it's considerably above average.}
 \item{3: }{I'm OK with this county's traditional cuisine. I think it's about average.}
 \item{2: }{I dislike this country's traditional cuisine. I think it's considerably below average.}
 \item{1: }{I hate this country's traditional cuisine. I think it's one of the worst in the world.}
 \item{N/A: }{I'm unfamiliar with this country's traditional cuisine.}
}
}
\seealso{
See \url{https://github.com/fivethirtyeight/data/tree/master/food-world-cup}
}
\keyword{datasets}

