% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meredith.R
\docType{data}
\name{generic_polllist}
\alias{generic_polllist}
\title{Congress Generic Ballot Polls}
\format{A data frame with 934 rows representing polls and 21 variables:
\describe{
  \item{subgroup}{No description provided.}
  \item{modeldate}{No description provided.}
  \item{startdate}{Start date of the poll.}
  \item{enddate}{End date of the poll.}
  \item{pollster}{The organization that conducted the poll (rather than the organization that paid for or sponsored it)}
  \item{grade}{No description provided.}
  \item{samplesize}{No description provided.}
  \item{population}{A = ALL ADULTS, RV = REGISTERED VOTERS, LV = LIKELY VOTERS, V = VOTERS}
  \item{weight}{No description provided.}
  \item{influence}{No description provided.}
  \item{dem}{No description provided.}
  \item{rep}{No description provided.}
  \item{adjusted_dem}{No description provided.}
  \item{adjusted_rep}{No description provided.}
  \item{multiversions}{No description provided.}
  \item{tracking}{No description provided.}
  \item{url}{No description provided.}
  \item{poll_id}{No description provided.}
  \item{question_id}{No description provided.}
  \item{createddate}{No description provided.}
  \item{timestamp}{No description provided.}
}}
\source{
See \url{https://github.com/fivethirtyeight/data/blob/master/congress-generic-ballot/README.md}
}
\usage{
generic_polllist
}
\description{
The raw data behind the story
"Are Democrats Winning The Race For Congress?"
\url{https://projects.fivethirtyeight.com/congress-generic-ballot-polls/}.
}
\seealso{
\code{\link{generic_topline}}
}
\keyword{datasets}
