% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{goose}
\alias{goose}
\title{The Save Ruined Relief Pitching. The Goose Egg Can Fix It.}
\format{A data frame with 30,533 rows representing pitchers and 12 variables:
\describe{
  \item{name}{Pitcher name}
  \item{year}{Start year of season}
  \item{team}{Retrosheet team code}
  \item{league}{NL or AL}
  \item{goose_eggs}{Goose eggs}
  \item{broken_eggs}{Broken eggs}
  \item{mehs}{Mehs}
  \item{league_average_gpct}{League-average goose percentage}
  \item{ppf}{Pitcher park factor}
  \item{replacement_gpct}{Replacement-level goose pecentage}
  \item{gwar}{Goose Wins Above Replacement}
  \item{key_retro}{Retrosheet unique player identifier}
}}
\source{
Retrosheet \url{http://www.retrosheet.org/}
}
\usage{
goose
}
\description{
The raw data behind the story
"The Save Ruined Relief Pitching. The Goose Egg Can Fix It."
\url{https://fivethirtyeight.com/features/goose-egg-new-save-stat-relief-pitchers/}.
}
\keyword{datasets}

