% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiphop_cand_lyrics}
\alias{hiphop_cand_lyrics}
\title{Hip-Hop Is Turning On Donald Trump}
\format{
A data frame with 377 rows representing hip-hop songs referencing POTUS candidates in 2016 and 8 variables:
\describe{
  \item{candidate}{Candidate referenced}
  \item{song}{Song name}
  \item{artist}{Artist name}
  \item{sentiment}{Positive, negative or neutral}
  \item{theme}{Theme of lyric}
  \item{album_release_date}{Date of album release}
  \item{line}{Lyrics}
  \item{url}{Genius link}
}
}
\source{
Genius https://genius.com/
}
\usage{
hiphop_cand_lyrics
}
\description{
The raw data behind the story
"Hip-Hop Is Turning On Donald Trump"
\url{https://projects.fivethirtyeight.com/clinton-trump-hip-hop-lyrics/}.
}
\keyword{datasets}
