% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{house_district_forecast}
\alias{house_district_forecast}
\title{2018 House Forecast}
\format{A dataframe with 302859 rows representing district-level results of the classic, lite, and deluxe   House forecasts since Aug. 1, 2018 and 11 variables
\describe{
\item{forecastdate}{date of the forecast}
\item{state}{state of the forecast}
\item{district}{district of the forecast}
\item{candidate}{name of the candidate}
\item{party}{party of the candidate}
\item{incumbent}{whether the candidate is incumbent}
\item{model}{the model of the forecast}
\item{win_probability}{the probability of the corresponding party winning}
\item{voteshare}{the voteshare of the corresponding party}
\item{p10_voteshare}{the top 10 percentile of the voteshare}
\item{p90_voteshare}{the top 00 percentile of the voteshare}
}}
\source{
FiveThirtyEight’s House, Senate And Governor Models
Methodology: 
\url{https://fivethirtyeight.com/methodology/how-fivethirtyeights-house-and-senate-models-work/}
}
\usage{
house_district_forecast
}
\description{
The raw data behind the story 
'Forecasting the race for the House'
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/house/}
}
\note{
the original dataset included an empty column "special", which was removed.
}
\keyword{datasets}
