% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fatima.R
\docType{data}
\name{impeachment_polls}
\alias{impeachment_polls}
\title{Do Americans Support Impeaching Trump?}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 388 rows and 24 columns.}
\source{
data from \url{https://github.com/fivethirtyeight/data/tree/master/impeachment-polls}.
}
\usage{
impeachment_polls
}
\description{
Raw data behind this story
"Do Americans Support Impeaching Trump?"
\url{https://projects.fivethirtyeight.com/impeachment-polls/}
}
\keyword{datasets}
