% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{mediacloud_hurricanes}
\alias{mediacloud_hurricanes}
\title{The Media Really Started Paying Attention to Puerto Rico When Trump Did}
\format{
A data frame with 38 rows representing dates and 5 variables:
\describe{
\item{date}{Date}
\item{harvey}{The number of sentences in online news which mention Hurricane Harvey on the specified date}
\item{irma}{The number of sentences in online news which mention Hurricane Irma}
\item{maria}{The number of sentences in online news which mention Hurricane Maria}
\item{jose}{The number of sentences in online news which mention Hurricane Jose}
}
}
\source{
Mediacloud \url{https://mediacloud.org/}
}
\usage{
mediacloud_hurricanes
}
\description{
The raw data behind the story
"The Media Really Started Paying Attention to Puerto Rico When Trump Did"
\url{https://fivethirtyeight.com/features/the-media-really-started-paying-attention-to-puerto-rico-when-trump-did/}:
Mediacloud Hurricanes Data.
}
\seealso{
\code{\link{mediacloud_states}}, \code{\link{mediacloud_online_news}},
\code{\link{mediacloud_trump}}, \code{\link{tv_hurricanes}}, \code{\link{tv_hurricanes_by_network}},
\code{\link{tv_states}}, \code{\link{google_trends}}
}
\keyword{datasets}
