% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{mediacloud_states}
\alias{mediacloud_states}
\title{Mediacloud States Data:}
\format{A data frame with 51 rows representing dates and 4 variables:
\describe{
\item{date}{Date}
\item{texas}{The number of sentences in online news which mention Texas on the specified date}
\item{puerto_rico}{The number of sentences in online news which mention Puerto Rico}
\item{florida}{The number of sentences in online news which mention Florida}
}}
\source{
Mediacloud \url{https://mediacloud.org/}
}
\usage{
mediacloud_states
}
\description{
Mediacloud States Data:
}
\seealso{
\code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_online_news}},
\code{\link{mediacloud_trump}}, \code{\link{tv_hurricanes}}, \code{\link{tv_hurricanes_by_network}},
\code{\link{tv_states}}, \code{\link{google_trends}}
}
\keyword{datasets}
