% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kara.R
\docType{data}
\name{nba_elo}
\alias{nba_elo}
\title{NBA Elo Ratings Through History}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 69636 rows and 24 columns.}
\usage{
nba_elo
}
\description{
The data behind the Elo and RAPTOR ratings for NBA teams used in various stories like this one: 
\url{https://projects.fivethirtyeight.com/complete-history-of-the-nba/#raptors}
}
\details{
@format A data frame with 69636 rows representing every game in the NBA, and 24 variables:
\describe{
  \item{date}{Date}
  \item{season}{the season in which the game was played}
  \item{neutral}{1 if the game was played on neutral territory, 0 if not}
  \item{playoff}{where in the playoff this game was played}
  \item{team1}{name of first team}
  \item{team2}{name of second team}
  \item{elo1_pre}{Team 1 Elo rating before game}
  \item{elo2_pre}{Team 2 Elo rating before game}
  \item{elo_prob1}{Team 1's probability of winning based on Elo rating}
  \item{elo_prob2}{Team 2's probability of winning based on Elo rating}
  \item{elo1_post}{Team 1 Elo rating after the game}
  \item{elo2_post}{Team 2 Elo rating after the game}
  \item{score1}{the score of team 1}
  \item{score2}{the score of team 2}
  }
}
\keyword{datasets}
