% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_starry.R
\docType{data}
\name{partisan_lean_state}
\alias{partisan_lean_state}
\title{FiveThirtyEight's Partisan Lean}
\format{A dataset with 50 rows representing states and 3 variables
\describe{
\item{state}{the state}
\item{pvi_party}{the party of the vote}
\item{pvi_amount}{the Cook Partisan Voting Index of the vote}
}}
\source{
Partisan lean is the average difference between how a state or district votes and how the country votes overall, with 2016 presidential election results weighted 50 percent, 2012 presidential election results weighted 25 percent and results from elections for the state legislature weighted 25 percent.
}
\usage{
partisan_lean_state
}
\description{
This directory contains the data for FiveThirtyEight's partisan lean, which is used in our
[House]
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/house}
[Senate]
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/senate}
and [Governer]
\url{https://projects.fivethirtyeight.com/2018-midterm-election-forecast/governor/}
forecasts.
}
\note{
The original dataset only has 2 columns: "state" and "pvi_538". I separated the "pvi_538" columns into two.
For example, in row 1 of the dataset, the original "pvi_538" = "R+27", and I separated it into “pvi_party" = "R" and "pvi_amount" = "27".
}
\seealso{
\code{\link{partisan_lean_district}}
}
\keyword{datasets}
