% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{pulitzer}
\alias{pulitzer}
\title{Do Pulitzers Help Newspapers Keep Readers?}
\format{A data frame with 50 rows representing newspapers and 7 variables:
\describe{
  \item{newspaper}{}
  \item{daily_circulation_2004}{}
  \item{daily_circulation_2013}{}
  \item{pctchange_in_daily_circulation_2004_2013}{}
  \item{winners_and_finalists_1990_2003}{}
  \item{winners_and_finalists_2004_2014}{}
  \item{winners_and_finalists_1990_2014}{}
}}
\source{
See \url{http://fivethirtyeight.com/datalab/do-pulitzers-help-newspapers-keep-readers/}
}
\usage{
pulitzer
}
\description{
The raw data behind the story
"Do Pulitzers Help Newspapers Keep Readers?"
\url{http://fivethirtyeight.com/datalab/do-pulitzers-help-newspapers-keep-readers/}.
}
\keyword{datasets}

