% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meredith.R
\docType{data}
\name{ratings}
\alias{ratings}
\title{An Inconvenient Sequel}
\format{A data frame with 80053 rows representing movie ratings and 27 variables:
\describe{
  \item{timestamp}{The date at which the rating was recorded.}
  \item{respondents}{The number of respondents in a category associated with a given timestamp.}
  \item{category}{The subgroups of respondents differentiated by demographics like gender, age, and nationality.}
  \item{link}{The website associated with a given category's responses.}
  \item{average}{The average rating reported by a given category.}
  \item{mean}{The mean rating reported by a given category.}
  \item{median}{The median rating reported by a given category.}
  \item{1_votes}{The count of votes denoting a rating of one that respondents gave.}
  \item{2_votes}{The count of votes denoting a rating of two that respondents gave.}
  \item{3_votes}{The count of votes denoting a rating of three that respondents gave.}
  \item{4_votes}{The count of votes denoting a rating of four that respondents gave.}
  \item{5_votes}{The count of votes denoting a rating of five that respondents gave.}
  \item{6_votes}{The count of votes denoting a rating of six that respondents gave.}
  \item{7_votes}{The count of votes denoting a rating of seven that respondents gave.}
  \item{8_votes}{The count of votes denoting a rating of eight that respondents gave.}
  \item{9_votes}{The count of votes denoting a rating of nine that respondents gave.}
  \item{10_votes}{The count of votes denoting a rating of ten that respondents gave.}
  \item{1_pct}{The percentage of votes denoting a rating of one that respondents gave.}
  \item{2_pct}{The percentage of votes denoting a rating of two that respondents gave.}
  \item{3_pct}{The percentage of votes denoting a rating of three that respondents gave.}
  \item{4_pct}{The percentage of votes denoting a rating of four that respondents gave.}
  \item{5_pct}{The percentage of votes denoting a rating of five that respondents gave.}
  \item{6_pct}{The percentage of votes denoting a rating of six that respondents gave.}
  \item{7_pct}{The percentage of votes denoting a rating of seven that respondents gave.}
  \item{8_pct}{The percentage of votes denoting a rating of eight that respondents gave.}
  \item{9_pct}{The percentage of votes denoting a rating of nine that respondents gave.}
  \item{10_pct}{The percentage of votes denoting a rating of ten that respondents gave.}
}}
\source{
IMBD \url{http://www.imdb.com/title/tt6322922/ratings} and 
see \url{https://github.com/fivethirtyeight/data/tree/master/inconvenient-sequel}
#'  @examples
# To convert data frame to tidy data (long) format, run:
library(tidyverse)
library(stringr)
ratings_tidy <- ratings %>%
  gather(votes, count, -c(timestamp, respondents, category, link, average, mean, median)) %>%
  arrange(timestamp)
}
\usage{
ratings
}
\description{
The raw data behind the story
"Al Gore’s New Movie Exposes The Big Flaw In Online Movie Ratings"
\url{https://fivethirtyeight.com/features/al-gores-new-movie-exposes-the-big-flaw-in-online-movie-ratings/}.
}
\keyword{datasets}
