% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trump_approval_trend}
\alias{trump_approval_trend}
\title{How Popular is Donald Trump}
\format{
A data frame with 1044 rows representing poll trends and 11 variables:
\describe{
\item{subgroup}{The subgroup the poll falls into as defined by the type of people being polled (all polls, voters, adults)}
\item{modeldate}{The date the model was created}
\item{approve_estimate}{Estimated approval ratings}
\item{approve_high}{Higher bound of the estimated approval percentage}
\item{approve_low}{Lower bound of the estimated approval percentage}
\item{disapprove_estimate}{Estimated disapproval percentage}
\item{disapprove_high}{Higher bound of the estimated disapproval percentage}
\item{disapprove_low}{Lower bound of the estimated disapproval percentage}
\item{timestamp}{Date and time the model was compiled}
}
}
\source{
\url{https://projects.fivethirtyeight.com/trump-approval-data/approval_topline.csv}
}
\usage{
trump_approval_trend
}
\description{
The raw data behind the story:
"How Popular is Donald Trump"
\url{https://projects.fivethirtyeight.com/trump-approval-ratings/}: Approval Trend Dataset.
}
\details{
The Variable "president" was removed because all values were "Donald Trump"
}
\seealso{
\code{\link{trump_approval_poll}}
}
\keyword{datasets}
