% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jen.R
\docType{data}
\name{trump_twitter}
\alias{trump_twitter}
\title{The World's Favorite Donald Trump Tweets}
\format{A data frame with 448 rows representing tweets and 3 variables:
\describe{
  \item{id}{}
  \item{created_at}{}
  \item{text}{}
}}
\source{
Twitter \url{https://twitter.com/realdonaldtrump}
}
\usage{
trump_twitter
}
\description{
The raw data behind the story
"The World's Favorite Donald Trump Tweets"
\url{https://fivethirtyeight.com/features/the-worlds-favorite-donald-trump-tweets/}.
Tweets posted on twitter by Donald Trump (@realDonaldTrump)
}
\keyword{datasets}
