% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{tv_hurricanes}
\alias{tv_hurricanes}
\title{TV Hurricanes Data:
 @format A data frame with 37 rows representing dates and 5 variables:
 \describe{
   \item{date}{Date}
   \item{harvey}{The percent of sentences in TV news that mention Hurricane Harvey on the given date}
   \item{irma}{The percent of sentences in TV news that mention Hurricane Irma}
   \item{maria}{The percent of sentences in TV news that mention Hurricane Maria}
   \item{jose}{The percent of sentences in TV news that mention Hurricane Irma}
 }
 @source Internet TV News Archive \url{https://archive.org/details/tv} and 
 Television Explorer \url{https://television.gdeltproject.org/cgi-bin/iatv_ftxtsearch/iatv_ftxtsearch}
 @seealso \code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_states}}, 
 \code{\link{mediacloud_online_news}}, \code{\link{mediacloud_trump}}, 
 \code{\link{tv_hurricanes_by_network}}, \code{\link{tv_states}}, \code{\link{google_trends}}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37 rows and 5 columns.}
\usage{
tv_hurricanes
}
\description{
TV Hurricanes Data:
 @format A data frame with 37 rows representing dates and 5 variables:
 \describe{
   \item{date}{Date}
   \item{harvey}{The percent of sentences in TV news that mention Hurricane Harvey on the given date}
   \item{irma}{The percent of sentences in TV news that mention Hurricane Irma}
   \item{maria}{The percent of sentences in TV news that mention Hurricane Maria}
   \item{jose}{The percent of sentences in TV news that mention Hurricane Irma}
 }
 @source Internet TV News Archive \url{https://archive.org/details/tv} and 
 Television Explorer \url{https://television.gdeltproject.org/cgi-bin/iatv_ftxtsearch/iatv_ftxtsearch}
 @seealso \code{\link{mediacloud_hurricanes}}, \code{\link{mediacloud_states}}, 
 \code{\link{mediacloud_online_news}}, \code{\link{mediacloud_trump}}, 
 \code{\link{tv_hurricanes_by_network}}, \code{\link{tv_states}}, \code{\link{google_trends}}
}
\keyword{datasets}
