% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_maggie.R
\docType{data}
\name{twitter_presidents}
\alias{twitter_presidents}
\title{The Worst Tweeter in Politics Isn't Trump}
\format{A data frame with 6439 rows describing individual tweets and 8 variables:
\describe{
\item{created_at}{The date and time the tweet was posted}
\item{user}{The user posting the tweet}
\item{text}{The text of the tweet}
\item{url}{The link to the tweet}
\item{replies}{The number of replies to the tweet}
\item{retweets}{The number of retweets}
\item{favorites}{The number of favorites}
}}
\source{
Twitter \url{https://twitter.com/BarackObama} and \url{https://twitter.com/realDonaldTrump}
}
\usage{
twitter_presidents
}
\description{
The raw data behind:
"The Worst Tweeter in Politics Isn't Trump"
\url{https://fivethirtyeight.com/features/the-worst-tweeter-in-politics-isnt-trump/}
}
\details{
Presidents Dataset:

Data on President Obama's tweets collected on Oct 20, President Trump's tweets collected on Oct 23.
}
\seealso{
\code{\link{senators}}
}
\keyword{datasets}
